% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sprawl_helpers.R
\name{areColors}
\alias{areColors}
\title{check if an array represents \code{R} colors}
\usage{
areColors(x)
}
\arguments{
\item{x}{\code{character array} containing color names}
}
\value{
\code{logical array} with value TRUE for entries correasponding to
colors, FALSE for invalid entries
}
\description{
check if all elements of a vector are valid "R" colors
}
\examples{
areColors(c("red", "green", "#C74C4C"))

areColors(c("red", "fdfd"))

}
\author{
Josh O'Brien - https://stackoverflow.com/a/13290832/6871135
}
