% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sprawl_helpers.R
\name{build_testraster}
\alias{build_testraster}
\title{build_testraster}
\usage{
build_testraster(nrows = 100, ncols = 100, nbands = 1,
  with_nodata = TRUE, tofile = FALSE)
}
\arguments{
\item{nrows}{\code{numeric} number of rows, Default: 100}

\item{ncols}{\code{numeric} number of columns, Default: 100}

\item{nbands}{\code{numeric} number of bands, Default: 1}

\item{with_nodata}{PARAM_DESCRIPTION, Default: TRUE}

\item{tofile}{\code{logical} if TRUE, the rasterStack is saved to a temporary file
and the name of the temporary filename is returned instead than the stack itself,
Default: FALSE}
}
\value{
either the "test" \code{rasterStack}, or a filename corresponding to it (if
\code{tofile} = TRUE)
}
\description{
FUNCTION_DESCRIPTION
}
\examples{
\dontrun{
if(interactive()){
 r <- build_testraster(100,100,10, tofile = F)
 }
}
}
