% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cast_vect.R
\name{cast_vect}
\alias{cast_vect}
\alias{cast_vect.default}
\alias{cast_vect.sf}
\alias{cast_vect.sfc}
\alias{cast_vect.Spatial}
\alias{cast_vect.character}
\title{cast a vector object to a different \code{R} spatial class or to a shapefile}
\usage{
cast_vect(object, to)

\method{cast_vect}{default}(object, to)

\method{cast_vect}{sf}(object, to)

\method{cast_vect}{sfc}(object, to)

\method{cast_vect}{Spatial}(object, to)

\method{cast_vect}{character}(object, to)
}
\arguments{
\item{object}{either an \code{R} object of class \code{sf}, \code{sfc} or \code{Spatial}, or a character string
corresponding to a filename (with full path)}

\item{to}{\code{character} indicating to which type of object the input should be re-casted. It can
be "sfobject" (re-cast to \code{sf}), "spobject" (recast to \code{Spatial}) or "vectfile" (re-cast to
a shapefile)}
}
\value{
returns the same object, casted to the "class" specified by \code{to} (or the exact same
object in case \code{object} is already of "class" \code{to})
}
\description{
function to automatically re-cast a "vector" object to a given "class" specified by
the \code{to} argument.
}
\details{
If \code{object} is a valid \code{R} spatial object, it is automatically converted to an object of
a different class if needed (e.g., from \code{sf} to \code{Spatial} and viceversa, or from \code{sf} to a
vector file through \code{sprawl::write_shape}). If it is a character string, the function checks if
it corresponds to a valid vector file and reads it to a \code{sf} or \code{Spatial} object through
\code{sprawl::read_vect}
}
\examples{
\dontrun{
 #EXAMPLE1
 }
}
\author{
Lorenzo Busetto, phD (2017) \href{mailto:lbusett@gmail.com}{lbusett@gmail.com}
}
