% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/change_fileext.R
\name{change_fileext}
\alias{change_fileext}
\title{change file extension}
\usage{
change_fileext(in_file, new_ext, new_path = NULL)
}
\arguments{
\item{in_file}{\code{character} full path of a file}

\item{new_ext}{\code{character} new file extension (e.g., ".png")}

\item{new_path}{\code{character} if not NULL the returned file name will use this
new path instead than the one of the original file. If equal to "", only
the updated basenmae is therefore returned (see examples), Default: NULL}
}
\value{
\code{character} basename/full_path of `in_file`` with updated extension
}
\description{
Helper function allowing to quckly change the extension of
a file name
}
\examples{
in_file = "home/lb/tmp/prova.tif"
# return an updated full_path
change_fileext(in_file, ".png")

#' update both path and extension
change_fileext(in_file, ".png", new_path = "/home/lb/new_path")

# return an updated basename
change_fileext(in_file, ".png", new_path = "")

}
\author{
Lorenzo Busetto, phD (2017) \href{mailto:lbusett@gmail.com}{lbusett@gmail.com}
}
