% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_spatype.R
\name{check_spatype}
\alias{check_spatype}
\title{check_spatype}
\usage{
check_spatype(object)
}
\arguments{
\item{object}{either a \code{R} object or a \code{character} string pointing to a vector or raster layer}
}
\value{
character vector equal to \emph{spobject}, \emph{sfobject}, \emph{rastobject}, \emph{vectfile} or \emph{rastfile}
}
\description{
accessory function to check if an object passed to the function corresponds to
a \code{*Spatial} Object, a \code{sf} object, a R \code{raster} object, a file corresponding to a vector,
or a file corresponding to a raster
}
\examples{
\dontrun{
# input is a raster file
check_spatype(system.file("extdata","testrast.tif", package = "sprawl"))

# input is a shapefile
shp_file <- system.file("extdata","lc_polys.shp", package = "sprawl")
check_spatype(shp_file)

# input is a `sp` object
obj <- readshape(shp_file, as_sp = TRUE)
check_spatype(obj)

# input is a `sf` object
obj <- readshape(shp_file)
check_spatype(obj)
}

}
