% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_vectype.R
\name{check_vectype}
\alias{check_vectype}
\title{check if an object is a \code{R} vector or vector file}
\usage{
check_vectype(in_vect, abort = FALSE)
}
\arguments{
\item{in_vect}{PARAM_DESCRIPTION}

\item{abort}{PARAM_DESCRIPTION}

\item{verbose}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
accessory function to check if an object passed to the function corresponds to
a \code{*sp} Object, a \code{sf} object, or a file corresponding to a valid ogr vector,
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[gdalUtils]{ogrinfo}}
}
