% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_rast_dtype.R
\name{convert_rastdtype}
\alias{convert_rastdtype}
\title{FUNCTION_TITLE}
\usage{
convert_rastdtype(dtype_string, type)
}
\arguments{
\item{dtype_string}{PARAM_DESCRIPTION}

\item{type}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\author{
Lorenzo Busetto, phD (2017) \href{mailto:lbusett@gmail.com}{lbusett@gmail.com}
}
