% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_virtrast.R
\name{create_virtrast}
\alias{create_virtrast}
\title{create a GDAL vrt file starting from a \code{Raster} object}
\usage{
create_virtrast(object, out_vrt_file, out_extent = NULL, rastinfo = NULL)
}
\arguments{
\item{out_vrt_file}{PARAM_DESCRIPTION}

\item{out_extent}{PARAM_DESCRIPTION, Default: NULL}

\item{rastinfo}{PARAM_DESCRIPTION, Default: NULL}

\item{in_file}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
create a GDAL vrt file starting from a \code{Raster} object
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\author{
Lorenzo Busetto, phD (2017) \href{mailto:lbusett@gmail.com}{lbusett@gmail.com}
}
