% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crop_rast.R
\name{crop_rast}
\alias{crop_rast}
\title{crop a raster file/object on a given extent}
\usage{
crop_rast(object, ext_object, out_filename = NULL, out_type = "rastobject",
  verbose = TRUE)
}
\arguments{
\item{object}{either an \code{R} object of class \code{Raster}, or a character string
corresponding to a raster filename (with full path)}

\item{extent}{\code{character} indicating to which type of object the input should be re-croped. It can
be "rastobject" (re-crop to \code{Raster}) or "rastfile" (recrop to raster file).In the
second case, recroping is achieved by identifying the raster file associated with the
Raster object and returning its path. If no file is associated (i.e., the object is
in-memory), the Raster object is written to disk on a temporary file, which path is returned}
}
\description{
function to crop a raster file/object on a given extent. The extent
can be:
\enumerate{
\item directly passed (as a \code{sprawlext} object or numeric array);
\item derived from a different object (raster of vector file/object) passed to
the function (see details)
}
}
\details{
If \code{object} is a valid \code{R} spatial object, it is automatically converted to an object of
a different class if needed (e.g., from \code{sf} to \code{Spatial} and viceversa, or from \code{sf} to a
vector file through \code{sprawl::write_shape}). If it is a character string, the function checks if
it corresponds to a valid raster file and reads it to a \code{Raster} object through
\code{rasterr::stack()}
}
\examples{
\dontrun{
 #EXAMPLE1
 }
}
\author{
Lorenzo Busetto, phD (2017) \href{mailto:lbusett@gmail.com}{lbusett@gmail.com}
}
