% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crop_vect.R
\name{crop_vect}
\alias{crop_vect}
\title{crop_vect}
\usage{
crop_vect(in_vect, in_obj, as_sp = FALSE)
}
\arguments{
\item{in_vect}{PARAM_DESCRIPTION}

\item{in_obj}{PARAM_DESCRIPTION}

\item{as_sp}{PARAM_DESCRIPTION, Default: FALSE}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[dplyr]{case_when}}
\code{\link[raster]{extent}},\code{\link[raster]{raster}}
\code{\link[sp]{crop}}
}
