% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cz_crop_object.R
\name{cz_crop_object}
\alias{cz_crop_object}
\title{cz_crop_object}
\usage{
cz_crop_object(in_vect_zones, in_rast, id_field, verbose)
}
\arguments{
\item{in_vect_zones}{PARAM_DESCRIPTION}

\item{in_rast}{PARAM_DESCRIPTION}

\item{id_field}{PARAM_DESCRIPTION}

\item{verbose}{PARAM_DESCRIPTION}
}
\value{
\code{list} with the following fields:
\itemize{
\item in_vect_zones_crop = cropped \code{sf} object
\item outside_feat = data frame containg info on the "removed" features (sequential id and corresponding
value ofthe "id" field (if provided)
}
}
\description{
Accessory function for \code{comp_zonal}. Crops the in_vect_zones on the extent of the raster
to improve speed. Then check if some features in the original in_vect_zones were dropped in
zone_raster_croppped
}
