% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cz_accessoires.R
\name{cz_getbands}
\alias{cz_getbands}
\title{cz_getbands}
\usage{
cz_getbands(in_rast, selbands = NULL, verbose = FALSE)
}
\arguments{
\item{in_rast}{PARAM_DESCRIPTION}

\item{selbands}{PARAM_DESCRIPTION, Default: NA}

\item{verbose}{PARAM_DESCRIPTION, Default: FALSE}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\examples{
\dontrun{
cz_getbands(in_rast, selbands = c(-1,050))
}
}
