% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cz_points.R
\name{cz_points}
\alias{cz_points}
\title{cz_points}
\usage{
cz_points(zone_object, in_rast, n_selbands, selbands, seldates, id_field,
  date_check, long, verbose, addfeat, addgeom, keep_null)
}
\arguments{
\item{zone_object}{\code{sf} object on which features the zonal statistics has to be computed}

\item{in_rast}{single- or multi-band \code{rasterStack} object containing values to be extracted and
summarized}

\item{n_selbands}{aa}

\item{selbands}{\code{2-element numeric array} defining starting and ending raster bands to be processed
(e.g., c(1, 10). Default: NULL, meaining that all bands will be processed}

\item{seldates}{aaa}

\item{id_field}{\code{character} (optional) field of the vector file to be used to identify the
zones from which data is to be extracted. MUST correspond to a univoc attribute (i.e., no replicated
values are allowed), If NULL (the default), zonal statistics are extracted on each row of the
shapefile, and a new column named \code{id_feat} is used on the output to distinguish the zones, Default: NULL}

\item{date_check}{zzz}

\item{long}{PARAM_DESCRIPTION, Default: FALSE}

\item{verbose}{PARAM_DESCRIPTION, Default: TRUE}

\item{addfeat}{PARAM_DESCRIPTION, Default: TRUE}

\item{addgeom}{PARAM_DESCRIPTION, Default: TRUE}

\item{keep_null}{PARAM_DESCRIPTION, Default: FALSE}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
