% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cz_poly.R
\name{cz_poly}
\alias{cz_poly}
\title{cz_poly}
\usage{
cz_poly(zone_object, in_rast, rastres, maxchunk, n_selbands, selbands, seldates,
  id_field, date_check, long, verbose, comp_quant, small, addfeat, addgeom,
  keep_null, zone_type, summ_data, full_data)
}
\arguments{
\item{zone_object}{PARAM_DESCRIPTION}

\item{in_rast}{PARAM_DESCRIPTION}

\item{rastres}{PARAM_DESCRIPTION, Default: NULL}

\item{maxchunk}{PARAM_DESCRIPTION, Default: 5e+07}

\item{n_selbands}{aa}

\item{selbands}{\code{2-element numeric array} defining starting and ending raster bands to be processed
(e.g., c(1, 10). Default: NULL, meaining that all bands will be processed}

\item{seldates}{aaa}

\item{id_field}{\code{character} (optional) field of the vector file to be used to identify the
zones from which data is to be extracted. MUST correspond to a univoc attribute (i.e., no replicated
values are allowed), If NULL (the default), zonal statistics are extracted on each row of the
shapefile, and a new column named \code{id_feat} is used on the output to distinguish the zones, Default: NULL}

\item{date_check}{zone_type}

\item{long}{PARAM_DESCRIPTION, Default: FALSE}

\item{verbose}{PARAM_DESCRIPTION, Default: TRUE}

\item{comp_quant}{\code{logical} if TRUE, also quantiles of the distributions of values are computed for each zone
and returned in  \code{out$ts_summ}, Default: FALSE}

\item{small}{\code{logical} if TRUE, values are returned also for small polygons not including any
pixel centroids, Default: TRUE}

\item{addfeat}{PARAM_DESCRIPTION, Default: TRUE}

\item{addgeom}{PARAM_DESCRIPTION, Default: TRUE}

\item{keep_null}{PARAM_DESCRIPTION, Default: FALSE}

\item{zone_type}{zone_type}

\item{summ_data}{\code{logical} If TRUE, summary statistics of values belonging to each vector zone
are returned in \code{out$ts_summ~. Default statistics are average, median, standard deviation and variation coefficient. Quantiles of the distribution can be also returned if}comp_quant == TRUE`,  Default: TRUE}

\item{full_data}{\code{logical} If TRUE, values of all pixels belonging to each vector zone are returned
in `out$ts_full', Default: TRUE}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
