% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cz_poly_parallel.R
\name{cz_poly_parallel}
\alias{cz_poly_parallel}
\title{FUNCTION_TITLE}
\usage{
cz_poly_parallel(zone_object, in_rast, rastres, maxchunk, n_selbands, selbands,
  seldates, id_field, date_check, long, verbose, comp_quant, small, addfeat,
  addgeom, keep_null, zone_type, summ_data, full_data, ncores)
}
\arguments{
\item{zone_object}{PARAM_DESCRIPTION}

\item{in_rast}{PARAM_DESCRIPTION}

\item{rastres}{PARAM_DESCRIPTION}

\item{maxchunk}{PARAM_DESCRIPTION}

\item{n_selbands}{PARAM_DESCRIPTION}

\item{selbands}{PARAM_DESCRIPTION}

\item{seldates}{PARAM_DESCRIPTION}

\item{id_field}{PARAM_DESCRIPTION}

\item{date_check}{PARAM_DESCRIPTION}

\item{long}{PARAM_DESCRIPTION}

\item{verbose}{PARAM_DESCRIPTION}

\item{comp_quant}{PARAM_DESCRIPTION}

\item{small}{PARAM_DESCRIPTION}

\item{addfeat}{PARAM_DESCRIPTION}

\item{addgeom}{PARAM_DESCRIPTION}

\item{keep_null}{PARAM_DESCRIPTION}

\item{zone_type}{PARAM_DESCRIPTION}

\item{summ_data}{PARAM_DESCRIPTION}

\item{full_data}{PARAM_DESCRIPTION}

\item{ncores}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[doSNOW]{registerDoSNOW}}
\code{\link[foreach]{foreach}}
\code{\link[gdalUtils]{gdal_rasterize}},\code{\link[gdalUtils]{gdalwarp}}
\code{\link[sp]{proj4string}}
\code{\link[tibble]{as_tibble}}
\code{\link[tidyr]{gather}}
\code{\link[utils]{txtProgressBar}},\code{\link[utils]{setTxtProgressBar}}
}
