% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datetodoy.R
\name{datetodoy}
\alias{datetodoy}
\title{Convert dates to DOYs}
\usage{
datetodoy(dates, abort = FALSE)
}
\arguments{
\item{dates}{array of dates to be converted. Class must be \code{Date} or character
parsable to \code{Date} using \code{\link{as.Date}} (format "yyyy-mm-dd").}

\item{abort}{\code{logical} If TRUE, the function aborts if even just one of the
dates fails converting (i.e., because one date is NA). If FALSE, a warning
is issued instead, Default: FALSE}
}
\description{
Converts an array of \code{Date} (or coercible to \code{Date}) values to
DOY (Day Of the Year) values. Simple wrapper around strftime(date, format = "\%j").
}
\examples{
 # Convert a date to a doy
 datetodoy(as.Date("2000-04-01"))

}
\author{
Lorenzo Busetto, PhD (2017) email: \href{mailto:lbusett@gmail.com}{lbusett@gmail.com}
}
