% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbReadSpatial.R
\name{dbReadSpatial}
\alias{dbReadSpatial}
\title{dbReadSpatial}
\usage{
dbReadSpatial(conn, name, pkey = NULL, ...)
}
\arguments{
\item{conn}{\code{DBIConnection} object, as produced by \code{\link[DBI:dbConnect]{DBI::dbConnect()}}.}

\item{name}{\code{character} the name of the spatial table / view / materialised view.
If the table is not in the 'public' schema, use a two-length character vector in the form
\code{c('schema_name','table_name')}.}

\item{pkey}{\code{character} (optional) the name of the primary key. This is needed only in the case of
views or materialised views (the primary key of a table is automatically retrieved).}

\item{...}{other options to be passed to....s}
}
\value{
An object of class \code{SpatialPixelsDataFrame}, \code{SpatialLinesDataFrame} or
\code{SpatialPolygonsDataFrame}, depending on the specific geometry which is being read.
}
\description{
Import a spatial table / view / materialised view from a \code{PostgreSQL}
database into a \code{Spatial*DataFrame} object.
}
\details{
The function reads a \code{PostGIS} table / view / materialised view from a \code{PostgreSQL} database,
in which a spatial column (format geom) must be present, and import it in \code{R} as a *\code{sp} object
\code{SpatialPixelsDataFrame}, \code{SpatialLinesDataFrame} or \code{SpatialPolygonsDataFrame}.
The name of the geometric column is automatically retrieved, while the primary key is retrieved
only reading a table (in case a view or materialised view is read, it must be specified with the
parameter \code{pkey}).
}
\note{
License: GPL 3.0
}
\examples{

}
\author{
Luigi Ranghetti, phD (2016) \email{ranghetti.l@irea.cnr.it}
}
