% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dissolve_shape.R
\name{dissolve_shape}
\alias{dissolve_shape}
\title{dissolve attributes of a vector}
\usage{
dissolve_shape(in_object, byvar, var_as_NA = FALSE)
}
\arguments{
\item{in_object}{PARAM_DESCRIPTION}

\item{byvar}{PARAM_DESCRIPTION}

\item{var_as_NA}{PARAM_DESCRIPTION, Default: FALSE}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
 library(sprawl)
 library(sprawl.data)
 indata    <- read_vect(system.file("extdata","lc_polys.shp", package = "sprawl.data"))
 byvar     <- "category"
 out_shape <- dissolve_shape(indata, byvar)
 out_shape
 }
}
\seealso{
\code{\link[tibble]{as_data_frame}}
}
