% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/er_getbands.R
\name{er_getbands}
\alias{er_getbands}
\title{er_getbands}
\usage{
er_getbands(in_rast, selbands = NULL, verbose = FALSE)
}
\arguments{
\item{in_rast}{PARAM_DESCRIPTION}

\item{selbands}{PARAM_DESCRIPTION, Default: NA}

\item{verbose}{PARAM_DESCRIPTION, Default: FALSE}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
