% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/er_points.R
\name{er_points}
\alias{er_points}
\title{helper for `extract_rast`` for points extraction}
\usage{
er_points(in_vect_crop, in_vect, in_rast, seldates, selbands, n_selbands,
  date_check, er_opts, outside_feat)
}
\arguments{
\item{in_vect_crop}{PARAM_DESCRIPTION}

\item{in_vect}{PARAM_DESCRIPTION}

\item{in_rast}{PARAM_DESCRIPTION}

\item{seldates}{PARAM_DESCRIPTION}

\item{selbands}{PARAM_DESCRIPTION}

\item{n_selbands}{PARAM_DESCRIPTION}

\item{date_check}{PARAM_DESCRIPTION}

\item{er_opts}{PARAM_DESCRIPTION}

\item{outside_feat}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\author{
Lorenzo Busetto, phD (2017) \href{mailto:lbusett@gmail.com}{lbusett@gmail.com}
}
