% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/er_polygons_std.R
\name{er_polygons_std}
\alias{er_polygons_std}
\title{extract raster data on polygons (helper for extract_rast)}
\usage{
er_polygons_std(in_vect_zones, in_rast, seldates, selbands, n_selbands,
  date_check, er_opts, verb_foreach = FALSE)
}
\arguments{
\item{in_vect_zones}{PARAM_DESCRIPTION}

\item{in_rast}{PARAM_DESCRIPTION}

\item{seldates}{PARAM_DESCRIPTION}

\item{selbands}{PARAM_DESCRIPTION}

\item{n_selbands}{PARAM_DESCRIPTION}

\item{date_check}{PARAM_DESCRIPTION}

\item{er_opts}{PARAM_DESCRIPTION}

\item{verb_foreach}{\code{logical} if TRUE, verbose output is sent out from within the foreach cycle,
Default: FALSE}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[doSNOW]{registerDoSNOW}}
\code{\link[foreach]{foreach}}
\code{\link[gdalUtils]{gdalwarp}}
\code{\link[sp]{proj4string}}
\code{\link[velox]{velox}}
\code{\link[utils]{setTxtProgressBar}},\code{\link[utils]{txtProgressBar}}
}
