% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extent_reproj.R
\name{extent_reproj}
\alias{extent_reproj}
\title{extent_reproj}
\usage{
extent_reproj(ext_in, in_proj = NULL, out_proj = NULL)
}
\arguments{
\item{ext_in}{object of class \code{\link[raster:extent]{raster::extent}}}

\item{in_proj}{\code{character} proj4string representing the projection of the input extent}

\item{out_proj}{\code{character} proj4string representing the desired projection for the output extent}
}
\value{
reprojected extent (in out_proj projection)
}
\description{
accessory function to convert the extent of a spatial object to a different
projection. simple wrapper for the \code{spTransform} function on an \code{\link[raster:extent]{raster::extent}} object
}
\examples{

}
