% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fast_summ.R
\name{fast_summ}
\alias{fast_summ}
\title{fast_summ}
\usage{
fast_summ(in_data, var, comp_quant, FUN, band_n, selband)
}
\arguments{
\item{in_data}{PARAM_DESCRIPTION}

\item{var}{PARAM_DESCRIPTION}

\item{comp_quant}{PARAM_DESCRIPTION}

\item{FUN}{PARAM_DESCRIPTION}

\item{band_n}{PARAM_DESCRIPTION}

\item{selband}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
