% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fillpals.R
\name{fillpals}
\alias{fillpals}
\title{create list of validi fill palettes}
\usage{
fillpals()
}
\value{
a \code{data.frame} containing info on valid fill palettes taht can used
in \code{plot_rast}, \code{plot_rast_gg} and \code{plot_vect}
}
\description{
helper function used to create a \code{data.frame} containing info
on valid fill palettes for \code{plot_rast}, \code{plot_rast_gg} and \code{plot_vect}
}
\details{
DETAILS
}
\examples{
\dontrun{
# Print the list of valid fill palettes used in `plot_rast`, `plot_rast_gg`
# and `plot_vect`
fillpals()
}
}
\author{
Lorenzo Busetto, phD (2017) \href{mailto:lbusett@gmail.com}{lbusett@gmail.com}
}
