% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_command.R
\name{find_command}
\alias{find_command}
\title{find the path to a specified command}
\usage{
find_command(command, stop_on_missing = TRUE)
}
\arguments{
\item{command}{\code{character} command to be search in the system (e.g., "gdalinfo")}

\item{stop_on_missing}{PARAM_DESCRIPTION, Default: TRUE}
}
\value{
\code{character} normalized path to the desired command
}
\description{
helper function allowing to find the path to a given command on
the system of the user. If the command is not found, a \code{stop} is issued and
the program aborts (gracefully).
}
\examples{

 # Find the path to `gdalinfo`
 gdalinfo_path <- find_command("gdalinfo")

}
\author{
Lorenzo Busetto \href{mailto:lbusett@gmail.com}{lbusett@gmail.com}
}
