% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_gdal.R
\name{find_gdal}
\alias{find_gdal}
\title{find main folder of gdal installation}
\usage{
find_gdal()
}
\value{
\code{character} full path to the gdal installation folder. The returned
object contains a \code{version} attribute that can be used to retrieve the
GDAL version. The function first checks if the path was already retrieved
and set in the "sprawl_gdalpath" global variable. If not, it tries to locate
the path using (in sequence):
1. Sys.which()
2. gdalUtils::gdal_setInstallation(ignore_fullscan = TRUE)
3. gdalUtils::gdal_setInstallation(ignore_fullscan = FALSE)

If all fails, the function aborts. If a valid installation is found, then
the global variable "sprawl_gdalpath" is set (this allows not to have
to continuously look for the gdal folder on later executions)
}
\description{
Helper function to identify the folder of the GDAL installation,
and its version
}
\examples{
 gdalpath <- find_gdal()
 gdalpath

}
\author{
Lorenzo Busetto, phD (2017) \href{mailto:lbusett@gmail.com}{lbusett@gmail.com}
}
