% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdal_polygonizeR.R
\name{gdal_polygonizeR}
\alias{gdal_polygonizeR}
\title{gdal_polygonizeR}
\usage{
gdal_polygonizeR(x, outshape = NULL, gdalformat = "ESRI Shapefile",
  pypath = NULL, readpoly = TRUE, quiet = TRUE, overwrite = FALSE)
}
\arguments{
\item{x}{\code{character}filename of a raster file, or "R" raster object}

\item{outshape}{\code{character}filename of the desired output polygon shapefile}

\item{gdalformat}{defaults to ESRI - don't change}

\item{pypath}{\code{character} path of python  - if \code{NULL} (the default) the script tries to
automatically retrieve it}

\item{readpoly}{\code{logical} If TRUE sends back the shapefile as a SpataialPolygonsDataFrame to R
(defaults to FALSE)}

\item{quiet}{\code{logical} if TRUE, limit the messages (defaults to TRUE)}

\item{overwrite}{\code{logical} If TRUE overwrite shapefile if already existing (defaults to FALSE)}
}
\value{
NULL, or a SpataialPolygonsDataFrame (if readpoly = TRUE)
}
\description{
R Wrapper for the gdal_polygonize.py python script (http://www.gdal.org/gdal_polygonize.html)
This utility creates vector polygons for all connected regions of pixels in the raster sharing
a common pixel value. Each polygon is created with an attribute indicating the pixel value of
that polygon. Can be userful for example to create fishnet polygons (see "create_fishnet")
}
\author{
Original code by Jonh Baumgartner
\url{https://johnbaumgartner.wordpress.com/2012/07/26/getting-rasters-into-shape-from-r/}, with
slight modifications by L.Busetto
}
