% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_extent.R
\name{get_extent}
\alias{get_extent}
\alias{get_extent.default}
\alias{get_extent.character}
\alias{get_extent.Raster}
\alias{get_extent.sf}
\alias{get_extent.sfc}
\alias{get_extent.Spatial}
\alias{get_extent.sprawlext}
\title{return the exten of a spatial object or file (with projection)}
\usage{
get_extent(object, abort = FALSE)

\method{get_extent}{default}(object, abort = FALSE)

\method{get_extent}{character}(object, abort = FALSE)

\method{get_extent}{Raster}(object, abort = FALSE)

\method{get_extent}{sf}(object, abort = FALSE)

\method{get_extent}{sfc}(object, abort = FALSE)

\method{get_extent}{Spatial}(object, abort = FALSE)

\method{get_extent}{Spatial}(object, abort = FALSE)
}
\arguments{
\item{object}{\code{character} corresponding to the name of a R object, or a filename
(full path)}

\item{abort}{\code{logical} if TRUE, the function aborts in case no proj4string or
invalid projstring is found, Default: FALSE}
}
\value{
object of class \code{sprawl_ext}
}
\description{
helper function used to retrieve the extent of a spatial object or file
with the associated proj4string
}
\details{
return an object of class \code{sprawl_ext} with two slots:
\itemize{
\item $extent: \code{numeric (4)} extent of the object (xmin, ymin, xmanx, ymax)
\item $projstring: \code{character} proj4string of the object
}
}
\examples{
\dontrun{
 library(raster)

 in_rast <- system.file("extdata", "sprawl_EVItest.tif", package = "sprawl.data")
 get_extent(in_rast)

 in_rast <- raster::raster(in_rast)
 get_extent(in_rast)

 in_vect <- system.file("extdata","lc_polys.shp", package = "sprawl.data")
 get_extent(in_vect)

 in_vect <- read_vect(in_vect)
 get_extent(in_vect)

}
}
\author{
Lorenzo Busetto, phD (2017) \href{mailto:lbusett@gmail.com}{lbusett@gmail.com}
}
