% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_PGIS.R
\name{get_PGIS}
\alias{get_PGIS}
\title{get_PGIS}
\usage{
get_PGIS(schema_name, table_name, con, geom_field = NULL, verbose = F)
}
\arguments{
\item{schema_name}{PARAM_DESCRIPTION}

\item{table_name}{PARAM_DESCRIPTION}

\item{con}{PARAM_DESCRIPTION}

\item{geom_field}{PARAM_DESCRIPTION, Default: NULL}

\item{verbose}{PARAM_DESCRIPTION, Default: F}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\examples{
 
}
\author{
Lorenzo Busetto, phD - Derived from: http://www.r-bloggers.com/load-postgis-geometries-in-r-without-rgdal
}
