% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_projstring.R
\name{get_projstring}
\alias{get_projstring}
\alias{get_projstring.default}
\alias{get_projstring.character}
\alias{get_projstring.Raster}
\alias{get_projstring.sf}
\alias{get_projstring.sfc}
\alias{get_projstring.Spatial}
\title{return the proj4string of a spatial object or file}
\usage{
get_projstring(object, abort = FALSE)

\method{get_projstring}{default}(object, abort = FALSE)

\method{get_projstring}{character}(object, abort = FALSE)

\method{get_projstring}{Raster}(object, abort = FALSE)

\method{get_projstring}{sf}(object, abort = FALSE)

\method{get_projstring}{sfc}(object, abort = FALSE)

\method{get_projstring}{Spatial}(object, abort = FALSE)
}
\arguments{
\item{object}{\code{character} corresponding to the name of a R object, or a filename
(full path)}

\item{abort}{\code{logical} if TRUE, the function aborts in case no proj4string or
invalid projstring is found, Default: FALSE}
}
\value{
\code{character} proj4string of the object or file
}
\description{
helper function used to extract the proj4string of "R" spatial objects
or of raster or vector files
}
\details{
DETAILS
}
\examples{
\dontrun{
 library(raster)

 in_rast <- system.file("extdata/MODIS_test", "EVIts_test.tif", package = "sprawl.data")
 get_projstring(in_rast)

 in_rast <- raster::raster(in_rast)
 get_projstring(in_rast)

 in_vect <- system.file("extdata/shapes","lc_polys.shp", package = "sprawl.data")
 get_projstring(in_vect)

 in_vect <- read_vect(in_vect)
 get_projstring(in_vect)

 }
}
\author{
Lorenzo Busetto, phD (2017) \href{mailto:lbusett@gmail.com}{lbusett@gmail.com}
}
