% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_projunits.R
\name{get_projunits}
\alias{get_projunits}
\title{Retrieve projection units}
\usage{
get_projunits(proj4string)
}
\arguments{
\item{proj4string}{valid proj4string or object of class \code{sp::CRS}}
}
\value{
\code{character} measure units of the provided proj4string (e.g., "m",
"dec.degrees", ...)
}
\description{
Function to retrieve projection units starting from a
valid proj4string or CRS
}
\examples{
\dontrun{
  # latlong projection
  proj_string <- "+init=epsg:4326"
  get_projunits(proj_string)

  # metric projection
  proj_string <- "+init=epsg:3857"
  get_projunits(proj_string)
}
}
\author{
Lorenzo Busetto, phD (2017) \href{mailto:lbusett@gmail.com}{lbusett@gmail.com}
}
