% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_rastinfo.R, R/get_raststats.R
\name{get_rastinfo}
\alias{get_rastinfo}
\alias{get_raststats}
\title{Retrieve useful info from a raster object or file}
\usage{
get_rastinfo(in_rast, stats = TRUE, quantiles = FALSE, hist = FALSE,
  verbose = TRUE)

get_raststats(in_rast, quantiles = FALSE, hist = FALSE)
}
\arguments{
\item{in_rast}{a \code{Raster} object or the filename of a valid raster}

\item{stats}{\code{logical} If TRUE, retrieve also statisics (min, max, mean, etc.)
for each band, Default: TRUE}

\item{quantiles}{\code{logical} If TRUE, retrieve also the quantiles, for each band,
Default; FALSE}

\item{hist}{\code{logical} If TRUE, retrieve also the frequency histogram, for
each band, Default: FALSE}

\item{verbose}{if FALSE suppress messages}

\item{in_rast}{a \code{Raster} object or the filename of a valid raster}

\item{quantiles}{\code{logical} If TRUE, retrieve also the quantiles, for each band,
Default; FALSE}

\item{hist}{\code{logical} If TRUE, retrieve also the frequency histogram, for
each band, Default: FALSE}

\item{verbose}{if FALSE suppress messages}
}
\value{
\code{list} containing the following object:
\itemize{
\item nbands: \code{numeric} Number of bands of the object/file;
\item indbands: \code{numeric} If the object is associated with a file on disk, bands of the
file to which each layer of the object correspond (see examples);
\item ncols: \code{numeric} Number of columns;
\item nrows: \code{numeric} Number of rows;
\item ncells: \code{numeric} Number of cells;
\item bnames: \code{character} Band names;
\item fnames: \code{character} If the object is associated with one or more files on disk,
filename(s) corresponding to the different bands;
\item Z: \code{ANY} if the object has a "z" attribute set using raster::setZ(), values of the
z attribute;
\item dtype: \code{character} datatype of the input raster (\code{raster} conventions)
\item proj4string: \code{character} proj4string of the object/file
\item units: \code{character} distance units of the projections
}

\code{list} containing the following object:
\itemize{
\item nbands: \code{numeric} Number of bands of the object/file;
\item indbands: \code{numeric} If the object is associated with a file on disk, bands of the
file to which each layer of the object correspond (see examples);
\item ncols: \code{numeric} Number of columns;
\item nrows: \code{numeric} Number of rows;
\item ncells: \code{numeric} Number of cells;
\item bnames: \code{character} Band names;
\item fnames: \code{character} If the object is associated with one or more files on disk,
filename(s) corresponding to the different bands;
\item Z: \code{ANY} if the object has a "z" attribute set using raster::setZ(), values of the
z attribute;
\item dtype: \code{character} datatype of the input raster (\code{raster} conventions)
\item proj4string: \code{character} proj4string of the object/file
\item units: \code{character} distance units of the projections
}
}
\description{
Function used to facilitate retrieval useful info from a raster object or
file.

Function used quickly retrieve statistics (min, max, mean, median
sd, and eventually quantiles) for bands of a raster object
}
\examples{
\dontrun{
 in_rast <- system.file("extdata/OLI_test", "oli_multi_1000.tif",
  package = "sprawl.data")
 get_rastinfo(in_rast)
 in_rast <- system.file("extdata/OLI_test", "oli_multi_1000_b1.tif",
  package = "sprawl.data")
 get_rastinfo(in_rast)
 }
\dontrun{
 in_rast <- system.file("extdata/OLI_test", "oli_multi_1000.tif",
  package = "sprawl.data")
 get_rastinfo(in_rast)
 in_rast <- system.file("extdata/OLI_test", "oli_multi_1000_b1.tif",
  package = "sprawl.data")
 get_rastinfo(in_rast)
 }
}
\author{
Lorenzo Busetto, phD (2017) \href{mailto:lbusett@gmail.com}{lbusett@gmail.com}

Lorenzo Busetto, phD (2017) \href{mailto:lbusett@gmail.com}{lbusett@gmail.com}
}
