% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_raststats.R
\name{get_raststats}
\alias{get_raststats}
\title{Retrieve statistics about value of a raster object}
\usage{
get_raststats(in_rast, quantiles = FALSE, hist = FALSE,
  verbose = TRUE)
}
\arguments{
\item{in_rast}{a \code{Raster} object or the filename of a valid raster}

\item{quantiles}{\code{logical} If TRUE, retrieve also the quantiles, for each band,
Default; FALSE}

\item{hist}{\code{logical} If TRUE, retrieve also the frequency histogram, for
each band, Default: FALSE}

\item{verbose}{if FALSE suppress messages, Default: TRUE}
}
\value{
a \code{list} containing the following elements:
\itemize{
\item \strong{stats}: \code{data.frame} containing min, max, average and standard deviation
for each band (see examples);
\item \strong{quants}: \code{data.frame} containing the quantiles of the distribution
of raster values, for each band (100 value - 0.01 interval) (NULL is returned
if \code{quants} == FALSE (the default));
\item \strong{hists}: \code{data.frame} containing information about the distribution of
raster values for each band. the data frame includes the limits of each
bin, the count of the number of pixels included in it, the corresponding
frequency and the cumulated frequency at each bin  (NULL is returned
if \code{hists} == FALSE (the default));
}
}
\description{
Function used quickly retrieve statistics (min, max, mean, median
sd, and eventually quantiles) for bands of a raster object
}
\examples{
 library(ggplot2)
 in_rast <- system.file("extdata/OLI_test", "oli_multi_1000.tif",
  package = "sprawl.data")
 get_raststats(in_rast)
 get_raststats(in_rast, quantiles = TRUE)
 get_raststats(in_rast, hist = TRUE, quantiles = TRUE)

 plot_rasthist(in_rast, type = "line") + xlim(-200,30000) + theme_bw()


}
\author{
Lorenzo Busetto, phD (2017) \href{mailto:lbusett@gmail.com}{lbusett@gmail.com}
}
