% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_rastype.R
\name{get_rastype}
\alias{get_rastype}
\alias{get_rastype.default}
\alias{get_rastype.character}
\alias{get_rastype.Raster}
\title{check the "spatial type" of an object or file}
\usage{
get_rastype(in_object)

\method{get_rastype}{default}(in_object)

\method{get_rastype}{character}(in_object)

\method{get_rastype}{Raster}(in_object)
}
\arguments{
\item{in_object}{either a \code{R} object or a \code{character} string pointing to a vector or raster layer}
}
\value{
character vector equal to \emph{spobject}, \emph{sfobject}, \emph{rastobject}, \emph{vectfile} or \emph{rastfile}
}
\description{
accessory function to check if an object passed to the function corresponds to
a \code{*Spatial} Object, a \code{sf} object, a R \code{raster} object, a file corresponding to a vector,
or a file corresponding to a raster
}
\examples{
\dontrun{
library(sprawl.data)
# input is a raster file
in_rast <- system.file("extdata/MODIS_test", "EVIts_test.tif", package = "sprawl.data")
get_rastype(in_rast)
}
}
