% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_spatype.R
\name{get_spatype}
\alias{get_spatype}
\alias{get_spatype.default}
\alias{get_spatype.character}
\alias{get_spatype.Raster}
\alias{get_spatype.sf}
\alias{get_spatype.sfc}
\alias{get_spatype.Spatial}
\title{check the "spatial type" of an object or file}
\usage{
get_spatype(object, abort = FALSE)

\method{get_spatype}{default}(object, abort = FALSE)

\method{get_spatype}{character}(object, abort = FALSE)

\method{get_spatype}{Raster}(object, abort = FALSE)

\method{get_spatype}{sf}(object, abort = FALSE)

\method{get_spatype}{sfc}(object, abort = FALSE)

\method{get_spatype}{Spatial}(object, abort = FALSE)
}
\arguments{
\item{object}{either a \code{R} object or a \code{character} string pointing to a vector or raster layer}

\item{abort}{\code{logical} if TRUE the function aborts if \code{object} is not recognized as an
R spatial file or valid vector or raster file}
}
\value{
character vector equal to \emph{spobject}, \emph{sfobject}, \emph{rastobject}, \emph{vectfile} or \emph{rastfile}
}
\description{
accessory function to check if an object passed to the function corresponds to
a \code{*Spatial} Object, a \code{sf} object, a R \code{raster} object, a file corresponding to a vector,
or a file corresponding to a raster
}
\examples{
\dontrun{
library(sprawl.data)
# input is a raster file
get_spatype(system.file("extdata","sprawl_EVItest.tif", package = "sprawl.data"))

# input is a shapefile
shp_file <- system.file("extdata","lc_polys.shp", package = "sprawl.data")
get_spatype(shp_file)

# input is a `sp` object
obj <- read_vect(shp_file, as_sp = TRUE)
get_spatype(obj)

# input is a `sf` object
obj <- read_vect(shp_file)
get_spatype(obj)
}

}
