% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_vectype.R
\name{get_vectype}
\alias{get_vectype}
\alias{get_vectype.default}
\alias{get_vectype.character}
\alias{get_vectype.sf}
\alias{get_vectype.sfc}
\alias{get_vectype.Spatial}
\alias{get_vectype.sprawlext}
\title{check the "spatial type" of a vector object or file}
\usage{
get_vectype(object)

\method{get_vectype}{default}(in_object)

\method{get_vectype}{character}(in_object)

\method{get_vectype}{sf}(object)

\method{get_vectype}{sfc}(object)

\method{get_vectype}{Spatial}(object)

\method{get_vectype}{Spatial}(object)
}
\arguments{
\item{object}{either a \code{R} object or a \code{character} string pointing to a vector
or raster layer}
}
\value{
character vector equal to \emph{spobject}, \emph{sfobject}, \emph{rastobject}, \emph{vectfile} or \emph{rastfile}
}
\description{
accessory function to check if an object passed to the function
corresponds to a \code{*Spatial} Object, a \code{sf} object, or a valid vector file
}
\examples{
\dontrun{
library(sprawl.data)

# input is a shapefile
in_vect <- system.file("extdata/shapes","lc_polys.shp", package = "sprawl.data")
get_vectype(in_vect)

# input is a `sp` object
obj <- read_vect(in_vect, as_sp = TRUE)
get_vectype(obj)

# input is a `sf` object
obj <- read_vect(in_vect)
get_vectype(obj)
}

}
