% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_scale_fill.R
\name{make_scale_fill}
\alias{make_scale_fill}
\title{FUNCTION_TITLE}
\usage{
make_scale_fill(palette, palette_type, na.color, zlims, leg_breaks, leg_labels,
  leg_type, outliers_style, direction)
}
\arguments{
\item{palette}{PARAM_DESCRIPTION}

\item{na.color}{PARAM_DESCRIPTION}

\item{zlims}{PARAM_DESCRIPTION}

\item{leg_breaks}{PARAM_DESCRIPTION}

\item{leg_labels}{PARAM_DESCRIPTION}

\item{leg_type}{PARAM_DESCRIPTION}

\item{outliers_style}{PARAM_DESCRIPTION}

\item{direction}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
#EXAMPLE1
 }
}
\author{
Lorenzo Busetto, phD (2017) \href{mailto:lbusett@gmail.com}{lbusett@gmail.com}
}
