% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maskrast.R
\name{maskrast}
\alias{maskrast}
\title{maskrast}
\usage{
maskrast(in_rast, mask_shape, buffer = NULL, verbose = TRUE)
}
\arguments{
\item{in_rast}{PARAM_DESCRIPTION}

\item{mask_shape}{PARAM_DESCRIPTION}

\item{buffer}{PARAM_DESCRIPTION, Default: NULL}

\item{verbose}{PARAM_DESCRIPTION, Default: TRUE}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[dplyr]{case_when}}
\code{\link[gdalUtils]{gdal_rasterize}}
\code{\link[raster]{raster}},\code{\link[raster]{origin}}
}
