% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/migrate_virtrast.R
\name{migrate_virtrast}
\alias{migrate_virtrast}
\title{migrate "virtual rasters" to a new location}
\usage{
migrate_virtrast(in_rast, newpath, out_file = NULL)
}
\arguments{
\item{in_rast}{\code{character} full file path of an \code{RData} or \code{vrt} file}

\item{newpath}{\code{character} new path were the corresponding files are now located. If NULL,
a file selector is opened allowing to interactively choose the new folder, Default: NULL}

\item{out_file}{\code{character} full path for the new \code{RData} or \code{vrt} file to be saved.
If NULL, the new file is built nby adding "_new" to the basename of the old one.
If == "overwrite", the old file is overwritten.  Default: NULL}
}
\value{
\code{character} path to the new "vrt" or "RData" file giving access to the raster
files
}
\description{
function allowing to update the paths contained in "virtual rasters"
such as GDAL VRTs or \code{R} rasterStacks "pointing" to files on disk in the case that
the corresponding files on disk are moved.
}
\details{
DETAILS
}
\examples{
\dontrun{
 # suppose you had a gdal vrt file "pointing" to tiff files originally located in
 # "/home/mypath/myfolder", and you successively moved them to "/home/mypath/mynewfolder"
 # and you want to update the vrt file so that it keeps working

 #TODO (Remove the # after uploading a test dataset on sprawl.data)
 #old_vrt  <- "/home/mypath/myfolder/myvrt.vrt
 #new_path <- "/home/mypath/mynewfolder"
 #new_vrt  <- migrate_virtrast(old_vrt, newpath)
 #new_vrt
 #raster::stack(new_vrt)
}
}
\author{
Lorenzo Busetto, phD (2017) \href{mailto:lbusett@gmail.com}{lbusett@gmail.com}
}
