% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pBuffer.R
\name{pBuffer}
\alias{pBuffer}
\title{pBuffer}
\usage{
pBuffer(points, res, res_scale = 1, quadsegs = 5, ...)
}
\arguments{
\item{points}{A \code{\linkS4class{SpatialPoints}} or \code{\linkS4class{SpatialPointsDataFrame}}
object, including the points to be buffered.}

\item{res}{A numeric vector of length 1 or 2, containing the radius of the circular buffers (if
1-length vector) or the width and height of the rectangle (if 2-length).
Alternatively, also a \code{\linkS4class{Raster}*} / \code{\linkS4class{SpatialGrid}*} /
\code{\linkS4class{SpatialPixels}*} is accepted: in this case, the raster resolution is used
as width/height of the rectangular buffers.}

\item{res_scale}{(optional) Numeric scaling factor of the resolution (useful in the case \code{res}
is a raster object).}

\item{quadsegs}{(optional) Number of line segments to use to approximate a quarter circle
(in case of circular buffer).}

\item{...}{other arguments}
}
\value{
An object of class \code{\linkS4class{SpatialPolygons}} or
\code{\linkS4class{SpatialPolygonsDataFrame}},
depending on the specific geometry which is read.
}
\description{
Create a \code{\linkS4class{SpatialPolygons}*} object expanding a
\code{\linkS4class{SpatialPoints}*} with single buffers. The new geometries are rectangles
or circles which maintain the original associated \code{data.frame}, if present.
The width/weight or radius of the buffers are directly provided to the function of taken from a sample
\code{\linkS4class{Raster}*} / \code{\linkS4class{SpatialGrid}*} / \code{\linkS4class{SpatialPixels}*}
object (see details).
}
\note{
License: GPL 3.0
}
\examples{
\dontrun{
# Load sample data
r <- raster::raster(system.file("extdata","testrast.tif", package = "sprawl"))
p <- sp::spsample(as(r, "SpatialPixels"), 20, "random")

# Rectangular buffer (3x3 cells)
pBuffer(p, r, 3)

# Squared buffer (100x100 metres)
pBuffer(p, c(100,100))

# Circular buffer (radius = 50 metres)
pBuffer(p, 50)
}

}
\author{
Luigi Ranghetti, phD (2017) \email{ranghetti.l@irea.cnr.it}
}
