% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_rast.R
\name{plot_rast}
\alias{plot_rast}
\title{FUNCTION_TITLE}
\usage{
plot_rast(in_rast, band = 1, in_poly = NULL, in_points = NULL,
  background = FALSE, limits = NULL, tails = c(0.02, 0.98),
  palette = "RdYlGn", legend_type = "standard", col_outlow = "gray10",
  col_outhigh = "gray90", maxpixels = 1e+05, title = "Raster Plot",
  plot_now = TRUE, ...)
}
\arguments{
\item{in_rast}{PARAM_DESCRIPTION}

\item{band}{PARAM_DESCRIPTION, Default: 1}

\item{in_poly}{PARAM_DESCRIPTION, Default: NULL}

\item{in_points}{PARAM_DESCRIPTION, Default: NULL}

\item{background}{PARAM_DESCRIPTION, Default: FALSE}

\item{limits}{PARAM_DESCRIPTION, Default: NULL}

\item{tails}{PARAM_DESCRIPTION, Default: c(0.01, 0.99)}

\item{palette}{PARAM_DESCRIPTION, Default: 'RdYlGn'}

\item{legend_type}{PARAM_DESCRIPTION, Default: 'standard'}

\item{col_outlow}{PARAM_DESCRIPTION, Default: 'gray10'}

\item{col_outhigh}{PARAM_DESCRIPTION, Default: 'gray90'}

\item{maxpixels}{PARAM_DESCRIPTION, Default: 1e+05}

\item{title}{PARAM_DESCRIPTION, Default: 'Raster Plot'}

\item{plot_now}{PARAM_DESCRIPTION, Default: TRUE}

\item{...}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 library(sprawl)

 in_rast <- system.file("extdata", "gNDVI.tif", package = "sprawl.data")
 in_poly <- create_fishnet(in_rast, pix_for_cell = 150)
 #'  # plot only the raster
 plot_rast(in_rast, in_poly = in_poly)

 # add a polygon and change the legend
 plot_rast(in_rast,
           in_poly   = in_vect,
           tails     = c(0.1, 0.99),
           legend    = "standard",
           palette   = "RdYlBu" ,
           maxpixels = 1e5)
 }
}
}
\seealso{
\code{\link[latticeExtra]{layer}}
\code{\link[raster]{stack}},\code{\link[raster]{quantile}}
\code{\link[rasterVis]{levelplot}}
\code{\link[RColorBrewer]{brewer.pal}}
}
