% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_rasthist.R
\name{plot_rasthist}
\alias{plot_rasthist}
\title{FUNCTION_TITLE}
\usage{
plot_rasthist(in_rast, variable = "freq", type = "hist", verbose = TRUE)
}
\arguments{
\item{in_rast}{PARAM_DESCRIPTION}

\item{variable}{PARAM_DESCRIPTION, Default: 'freq'}

\item{type}{PARAM_DESCRIPTION, Default: 'hist'}

\item{verbose}{PARAM_DESCRIPTION, Default: TRUE}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
#EXAMPLE1
}
\author{
Lorenzo Busetto, phD (2017) \href{mailto:lbusett@gmail.com}{lbusett@gmail.com}
}
