% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_vect.R
\name{plot_vect}
\alias{plot_vect}
\title{plot a vector}
\usage{
plot_vect(in_data, fill_var, borders, xlims = NULL, ylims = NULL,
  leg_type = NULL, leg_labels = NULL, leg_breaks = NULL,
  no_axis = FALSE, title = "Vector Plot", subtitle = NULL,
  scalebar = TRUE, scalebar_dist = NULL, palette_type = "diverging",
  palette_name = NULL, direction = 1, theme = theme_bw(),
  verbose = TRUE)
}
\arguments{
\item{in_data}{PARAM_DESCRIPTION}

\item{fill_var}{PARAM_DESCRIPTION}

\item{borders}{PARAM_DESCRIPTION}

\item{xlims}{PARAM_DESCRIPTION, Default: NULL}

\item{ylims}{PARAM_DESCRIPTION, Default: NULL}

\item{leg_type}{PARAM_DESCRIPTION, Default: NULL}

\item{leg_labels}{PARAM_DESCRIPTION, Default: NULL}

\item{leg_breaks}{PARAM_DESCRIPTION, Default: NULL}

\item{no_axis}{PARAM_DESCRIPTION, Default: FALSE}

\item{title}{PARAM_DESCRIPTION, Default: 'Vector Plot'}

\item{subtitle}{PARAM_DESCRIPTION, Default: NULL}

\item{scalebar}{PARAM_DESCRIPTION, Default: TRUE}

\item{scalebar_dist}{PARAM_DESCRIPTION, Default: NULL}

\item{palette_type}{PARAM_DESCRIPTION, Default: 'diverging'}

\item{palette_name}{PARAM_DESCRIPTION, Default: NULL}

\item{direction}{PARAM_DESCRIPTION, Default: 1}

\item{theme}{PARAM_DESCRIPTION, Default: theme_bw()}

\item{verbose}{PARAM_DESCRIPTION, Default: TRUE}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\author{
Lorenzo Busetto, phD (2017) \href{mailto:lbusett@gmail.com}{lbusett@gmail.com}
}
