% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readshape.R
\name{readshape}
\alias{readshape}
\title{readshape}
\usage{
readshape(shp_file, as_sp = FALSE, ...)
}
\arguments{
\item{shp_file}{\code{character} Filename of ESRI shapefile to be opened}

\item{as_sp}{\code{logical} If TRUE, the opened object is automatically converted to \code{sp} format.
Otherwise, an \code{sf} object is returned. Default: FALSE}

\item{...}{other arguments to be passed to\link[sf:st_read]{sf::st_read}}
}
\value{
Spatial* object (SpatialPolygons, SpatialPoints, eccetera)
}
\description{
function for easily opening a ESRI shapefile (or other OGR compatible vector file)
by simply specifying its filename
}
\examples{
\dontrun{
# open a shapefile as a `sf` object
 shp_file = system.file("extdata","clip_shape.shp", package = "sprawl")
 readshape(shp_file)

# open a shapefile as a `sp` object
 shp_file = system.file("extdata","clip_shape.shp", package = "sprawl")
 readshape(shp_file, as_sp = TRUE)
}
}
