% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reclass_rast.R
\name{reclass_rast}
\alias{reclass_rast}
\title{reclass_rast}
\usage{
reclass_rast(in_rast, rcl_mat, out_rast, r_out = FALSE, ovr = FALSE)
}
\arguments{
\item{in_rast}{Input raster file or "R" raster layer to be reclassified}

\item{rcl_mat}{input reclassification matrix. list of list created  as follows:
rcl_mat <- list(
list(start = X,  end  = Y , new = n1),
list(start = X1, end  = Y1, new = n2),
.....,
list(start = Xn, end = Yn, new = nn)
)
IMPORTANT ! Intervals are CLOSED on THE LEFT and OPEN ON THE RIGHT !
\itemize{
\item see examples !!!! -
}}

\item{out_rast}{name of the output raster file}

\item{r_out}{logical flag indicating if the output raster should be sent back to "R"}

\item{ovr}{logical indicateing if the output file should be overwritten if already existing}
}
\value{
if (r_out == TRUE), returns a "R" raster object corresponding to the reclassified raster
}
\description{
reclass_rast
}
\examples{
\dontrun{
# reclassify a land cover map with N classes to a 0-1 mask, retaining classes 5 and 9, putting
# the rest to 0 and values >= 11 to NA
# Open  the masking file
in_mask <- raster(in_maskfile)
# setup the reclassification matrix
rcl_mat <- list(
   list(start = 0, end  = 0, new = NA),
   list(start = 1, end  = 5, new = 0),
   list(start = 5, end  = 6, new = 1),
   list(start = 6, end  = 9, new = 1),
   list(start = 9, end  = 11, new = 1),
   list(start = 11, end = 100, new = NA)
  )


reclass_file = "/home/lb/Temp/buttami/pippo_reclass.tif"
outmask = raster_reclassifier(in_mask, rcl_mat, out_rast = reclass_file,
                                r_out = TRUE, ovr = TRUE)
summary(outmask)
}
}
