% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reproj_rast.R
\name{reproj_rast}
\alias{reproj_rast}
\title{reproj_rast}
\usage{
reproj_rast(object, out_proj, out_extent = NULL, resamp_meth = "near",
  out_type = "rastobject", out_format = "GTiff", compression = "LZW",
  out_filename = NULL, overwrite = FALSE, verbose = TRUE, ...)
}
\arguments{
\item{object}{PARAM_DESCRIPTION}

\item{out_proj}{PARAM_DESCRIPTION}

\item{out_extent}{PARAM_DESCRIPTION, Default: NULL}

\item{resamp_meth}{PARAM_DESCRIPTION, Default: 'near'}

\item{out_type}{PARAM_DESCRIPTION, Default: 'rastobject'}

\item{out_format}{PARAM_DESCRIPTION, Default: 'Gtiff'}

\item{compression}{PARAM_DESCRIPTION, Default: 'LZW'}

\item{out_filename}{PARAM_DESCRIPTION, Default: NULL}

\item{overwrite}{PARAM_DESCRIPTION, Default: FALSE}

\item{verbose}{PARAM_DESCRIPTION, Default: TRUE}

\item{...}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 in_rast <- read_rast(system.file("extdata/OLI_test",
  "oli_multi_1000_b2.tif", package = "sprawl.data"))
 out_proj <- "+init=epsg=3035"
 reproj_rast(in_rast, out_proj)
 }
}
}
\author{
Lorenzo Busetto, phD (2017) \href{mailto:lbusett@gmail.com}{lbusett@gmail.com}
}
