% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_extent.R
\name{select_extent}
\alias{select_extent}
\title{FUNCTION_TITLE}
\usage{
select_extent(in_object = NULL, transparency = 0.5)
}
\arguments{
\item{in_object}{PARAM_DESCRIPTION, Default: NULL}

\item{transparency}{PARAM_DESCRIPTION, Default: 0.5}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
if(interactive()){
 extent    <- select_extent()
 }

}
\author{
Lorenzo Busetto, phD (2017) \href{mailto:lbusett@gmail.com}{lbusett@gmail.com}
}
