% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_rastlabels.R
\name{set_rastlabels}
\alias{set_rastlabels}
\title{categorize a raster and assign class names}
\usage{
set_rastlabels(in_rast, class_names = NULL, verbose = TRUE)
}
\arguments{
\item{in_rast}{a \code{Raster} object or the path to a valid rster file}

\item{class_names}{\code{character array} optional array of names to be assigned
to the new classes (useful for plotting with \code{plot_rast} or \code{plot_rast_gg}) -
see examples. Order of the class names should follow the numeric sortuing
of the values to which class names should correspond. If NULL, class names
are assigned by converting the numeric values of the new raster to character,
Default: NULL}

\item{verbose}{\code{logical} if TRUE, extended processing information is sent to
the console in the form of messages}
}
\value{
a 'Raster' object equal to \code{in_rast} to which a Raster Attribute table
has been added (see \code{raster::ratify})
}
\description{
FUNCTION_DESCRIPTION
}
\details{
Simple wrapper around \code{raster::ratify}, providing the added
functionality of specifying specific names for the classes.
}
\examples{
\dontrun{
  in_rast <- raster::raster(ncol = 5, nrow = 5) \%>\%
    init("row")
  levels(in_rast)[[1]]

  # categorize and assign standard class names
  cat_rast <- set_rastlabels(in_rast)
  levels(cat_rast)[[1]]

  # categorize and assign custom class names
  cat_rast <- set_rastlabels(in_rast,
                              class_names = c("Water", "Land",
                              "Vegetation", "Urban", "Other"))
  levels(cat_rast)[[1]]
  plot_rast_gg(cat_rast, leg_labels = levels(cat_rast)[[1]]$Class,
               scalebar = FALSE)
 }
}
\seealso{
\code{raster::ratify}
}
\author{
Lorenzo Busetto, phD (2017) \href{mailto:lbusett@gmail.com}{lbusett@gmail.com}
}
