% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{setClasses}
\alias{setClasses}
\title{FUNCTION_TITLE}
\usage{
setClasses()
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION

An S4 class to represent the extent of a spatial object, associated with its proj4
string.
}
\details{
DETAILS
}
\section{Slots}{

\describe{
\item{\code{extent}}{\code{numeric (4)} extent of the object (xmin, ymin, xmanx, ymax)}

\item{\code{projstring}}{character` proj4string of the object}
}}

