% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sieve_rast.R
\name{sieve_rast}
\alias{sieve_rast}
\title{FUNCTION_TITLE}
\usage{
sieve_rast(in_rast, min_patch_area, out_type = "rastobject",
  out_file = NULL, overwrite = FALSE, verbose = TRUE)
}
\arguments{
\item{in_rast}{PARAM_DESCRIPTION}

\item{min_patch_area}{PARAM_DESCRIPTION}

\item{out_type}{PARAM_DESCRIPTION}

\item{out_file}{PARAM_DESCRIPTION}

\item{overwrite}{PARAM_DESCRIPTION}

\item{verbose}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\author{
Lorenzo Busetto, phD (2017) \href{mailto:lbusett@gmail.com}{lbusett@gmail.com}
}
