% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sieve_vect.R
\name{sieve_vect}
\alias{sieve_vect}
\title{FUNCTION_TITLE}
\usage{
sieve_vect(in_vect, min_patch_area, class_field, reassign_met = "bord_lgt",
  out_type = "sfobject", out_file = NULL, overwrite = FALSE,
  create_dir = FALSE, verbose = TRUE)
}
\arguments{
\item{in_vect}{PARAM_DESCRIPTION}

\item{min_patch_area}{PARAM_DESCRIPTION}

\item{reassign_met}{PARAM_DESCRIPTION}

\item{out_type}{PARAM_DESCRIPTION, Default: NULL}

\item{out_file}{PARAM_DESCRIPTION, Default: NULL}

\item{overwrite}{PARAM_DESCRIPTION, Default: NULL}

\item{create_dir}{PARAM_DESCRIPTION, Default: NULL}

\item{verbose}{PARAM_DESCRIPTION, Default: NULL}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\author{
Lorenzo Busetto, phD (2017) \href{mailto:lbusett@gmail.com}{lbusett@gmail.com}
}
