% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sprawl_initcluster.R
\name{sprawl_initcluster}
\alias{sprawl_initcluster}
\title{initialize clustered processing on multiband raster}
\usage{
sprawl_initcluster(in_rast, ncores = NULL, bands = NULL, maxchunk = 5e+06,
  out_on_console = FALSE)
}
\arguments{
\item{in_rast}{PARAM_DESCRIPTION}

\item{ncores}{PARAM_DESCRIPTION, Default: NULL}

\item{bands}{PARAM_DESCRIPTION, Default: NULL}

\item{maxchunk}{PARAM_DESCRIPTION, Default: 5e+05}

\item{out_on_console}{PARAM_DESCRIPTION, Default: FALSE}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
helper for initializing clustered processing on multiband raster
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\author{
Lorenzo Busetto, phD (2017) \href{mailto:lbusett@gmail.com}{lbusett@gmail.com}
}
