% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sprawlext-class.R
\name{sprawlext}
\alias{sprawlext}
\title{sprawlext class}
\description{
An S4 class to represent the extent of a spatial object,
associated with its proj4string.
}
\section{Slots}{

\describe{
\item{\code{extent}}{\code{numeric (4)} extent of the object (xmin, ymin, xmax, ymax)}

\item{\code{proj4string}}{\code{character} proj4string of the object}
}}

\examples{
showClass("sprawlext")

\dontrun{

library(raster)

in_rast <- system.file("extdata/MODIS_test", "EVIts_test.tif",
  package = "sprawl.data")
ex_sprawlext <- get_extent(in_rast)

# Convert to Extent
ex_extent <- as(ex_sprawlext, "Extent")
ex_extent

# Convert to bbox
ex_bbox <- as(ex_sprawlext, "matrix")
ex_bbox

# Convert to Spatial* objects
as(ex_sprawlext, "SpatialPolygons")
as(ex_sprawlext, "SpatialPoints")
as(ex_sprawlext, "SpatialLines")

# Convert to sf objects
as(ex_sprawlext, "st_polygon")
as(ex_sprawlext, "st_multipoint")

# Extract proj4string CRS
ex_crs <- as(ex_sprawlext, "CRS")
ex_crs

# Create sprawlext from extent and CRS
get_extent(ex_extent, ex_crs)

# Create sprawlext from bbox and CRS
get_extent(ex_bbox, ex_crs)
}
}
