% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_data.R
\name{summarize_data}
\alias{summarize_data}
\title{helper function used to extract statistics of raster values within
\code{sprawl::extract_rast} and \code{sprawl::aggregate_rast}}
\usage{
summarize_data(in_data, var, comp_quant, FUN, band_n, selband)
}
\arguments{
\item{in_data}{\code{data.table} containing raster values (for a full file or a "chunk")}

\item{var}{\code{character} name of the variable (i.e., column) to be used as grouping
variable for summarization}

\item{comp_quant}{\code{logical} if TRUE, percentiles of the distribution are computed
alongside standard statistics}

\item{FUN}{\code{function} if not NULL and is  a valid function, FUN is used for summarization
and only its results are reported. Useful for example for analyses over large
high-res raster since it reduces memory footprint (see \code{sprawl::aggregate_rast})}

\item{band_n}{PARAM_DESCRIPTION}

\item{selband}{PARAM_DESCRIPTION}
}
\value{
\code{list}
}
\description{
FUNCTION_DESCRIPTION
}
