% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_shape.R
\name{write_shape}
\alias{write_shape}
\title{Write a \code{sf} or \code{sp} to a ESRI shapefile}
\usage{
write_shape(in_vect, out_file, overwrite = FALSE, verbose = FALSE,
  encoding = "UTF-8", create_dir = FALSE, ...)
}
\arguments{
\item{in_vect}{\code{character} object to be written to the shapefile. Must be a
valid \code{*sp} or \code{sf} object}

\item{out_file}{\code{character} output file name}

\item{overwrite}{\code{logical} if TRUE, output file will be overwritten if existing,
Default: FALSE}

\item{verbose}{\code{logical} if TRUE, provide messages on processing, Default: TRUE}

\item{encoding}{\code{character} encoding to be used to write the DBF of the
shapefile, Default: "UTF-8"}

\item{create_dir}{\code{logical}, if TRUE and the folder of out_file does not exist,
then the full folder tree up to out_file is created, Default: FALSE
(use with caution !)}

\item{...}{any other arguments to be passed to \code{sf::st_write}}
}
\description{
Wrapper on  \code{sf::st_write} for writing an \code{sf} or \code{sp} object to
disk as an ESRI shapefile
}
\examples{

# build a spatialpoints data frame
pts                <- cbind(1:5, 1:5)
dimnames(pts)[[1]] <- letters[1:5]
df                 <-  data.frame(a = 1:5)
row.names(df)      <- letters[5:1]
points             <- sp::SpatialPointsDataFrame(pts, df, match.ID = TRUE)

# Save it to a shape file
out_file  = tempfile(pattern = "test", tmpdir = tempdir(),
                         fileext = ".shp")
mysp_object = points
write_shape(mysp_object, out_file, overwrite = TRUE)

# read the saved shape
in_data <- read_vect(out_file)
in_data


}
