% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_vector_to_pgres.R
\name{add_vector_to_pgres}
\alias{add_vector_to_pgres}
\title{write spatial table to a PostGIS db table}
\usage{
add_vector_to_pgres(vector_to_add, con, table_name, schema = "public",
  drop = FALSE)
}
\arguments{
\item{vector_to_add}{either an "sf" or "sp" R object, or the full path of
a valid vector file}

\item{con}{a \code{PostgreSQLConnection} object to a PostGIS DB, created by a call to
\code{DBI::DBI::dbConnect(PostgreSQL(),....}}

\item{table_name}{\code{character} name of the table to be used for storage in the db}

\item{schema}{\code{character} name of the schema to be used for storage in the db}

\item{drop}{\code{logical} indicating if existing tables should be overwrittern,
Default: FALSE}
}
\value{
The function is called for its side effects
}
\description{
write spatial table to a PostGIS db table
}
\details{
DETAILS
}
\examples{
\dontrun{
con <- DBI::dbConnect(PostgreSQL(),
                     host = "xx.xxx.xx.xx", dbname = "xxxxx", user = "xxxxx",
                     password = "xxxxx", port = 5432)
> con
<PostgreSQLConnection>

 write_vector_to_pgres(in_vect,
                  con        = con,
                  table_name = "my_table",
                  schema     = "public",
                  drop       = TRUE)
}
}
\author{
Lorenzo Busetto, phD (2017) \href{mailto:lbusett@gmail.com}{lbusett@gmail.com}
}
