% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_scale_fill.R
\name{add_scale_fill}
\alias{add_scale_fill}
\title{FUNCTION_TITLE}
\usage{
add_scale_fill(plot, palette, title, na.color, zlims, leg_breaks, leg_labels,
  leg_type, outliers_style, direction, ...)
}
\arguments{
\item{plot}{PARAM_DESCRIPTION}

\item{palette}{PARAM_DESCRIPTION}

\item{title}{PARAM_DESCRIPTION}

\item{na.color}{PARAM_DESCRIPTION}

\item{zlims}{PARAM_DESCRIPTION}

\item{leg_breaks}{PARAM_DESCRIPTION}

\item{leg_labels}{PARAM_DESCRIPTION}

\item{leg_type}{PARAM_DESCRIPTION}

\item{outliers_style}{PARAM_DESCRIPTION}

\item{direction}{PARAM_DESCRIPTION}

\item{...}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\author{
Lorenzo Busetto, phD (2017) \href{mailto:lbusett@gmail.com}{lbusett@gmail.com}
}
