% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate_rast.R
\name{aggregate_rast}
\alias{aggregate_rast}
\title{aggregate_rast}
\usage{
aggregate_rast(in_rast_values, zones_rast, FUN = mean, maxchunk = 5e+07,
  method = "fastdisk", to_file = FALSE, out_file = NULL,
  nodata_out = NA, verbose = TRUE)
}
\arguments{
\item{in_rast_values}{PARAM_DESCRIPTION}

\item{zones_rast}{PARAM_DESCRIPTION}

\item{FUN}{PARAM_DESCRIPTION, Default: mean}

\item{maxchunk}{PARAM_DESCRIPTION, Default: 5e+07}

\item{method}{PARAM_DESCRIPTION, Default: 'fastdisk'}

\item{to_file}{PARAM_DESCRIPTION, Default: FALSE}

\item{out_file}{PARAM_DESCRIPTION, Default: NULL}

\item{nodata_out}{PARAM_DESCRIPTION, Default: NA}

\item{verbose}{PARAM_DESCRIPTION, Default: TRUE}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
Function used to extract values of a single- or multi- band raster for pixels
corresponding to the features of a vector (Polygon, Point or Line)
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[data.table]{data.table}}
\code{\link[raster]{raster}},\code{\link[raster]{writeRaster}},\code{\link[raster]{rasterize}}
\code{\link[sf]{st_crs}},\code{\link[sf]{st_transform}},\code{\link[sf]{st_bbox}},\code{\link[sf]{st_coordinates}},\code{\link[sf]{st_centroid}},\code{\link[sf]{st_as_sf}},\code{\link[sf]{st_set_crs}}
\code{\link[sp]{proj4string}}
}
\author{
Lorenzo Busetto, PhD (2017) email: \href{mailto:lbusett@gmail.com}{lbusett@gmail.com}
}
