% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate_rast.R
\name{aggregate_rast}
\alias{aggregate_rast}
\title{aggregate_rast}
\usage{
aggregate_rast(in_val_rast, in_zones_rast, FUN = mean, maxchunk = 5e+07,
  method = "disk", out_type = "rastobject", out_file = NULL,
  nodata_in = NULL, verbose = TRUE, overwrite = FALSE)
}
\arguments{
\item{in_val_rast}{Object of class \code{Raster*} or filename corresponding to a
valid raster file}

\item{in_zones_rast}{PARAM_DESCRIPTION}

\item{FUN}{PARAM_DESCRIPTION, Default: mean}

\item{maxchunk}{PARAM_DESCRIPTION, Default: 5e+07}

\item{method}{PARAM_DESCRIPTION, Default: "disk"}

\item{out_type}{PARAM_DESCRIPTION, Default: 'rastobject'}

\item{out_file}{PARAM_DESCRIPTION, Default: NULL}

\item{nodata_in}{PARAM_DESCRIPTION, Default: NULL}

\item{verbose}{PARAM_DESCRIPTION, Default: TRUE}

\item{overwrite}{\code{logical} If TRUE, existing files are overwritten,
Default: FALSE}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
Function used to extract values of a single- or multi- band
raster for pixels corresponding to the features of a vector (Polygon,
Point or Line)
}
\details{
DETAILS
}
\examples{
 in_file <- system.file("extdata/MODIS_test", "EVIts_test.tif",
                                         package = "sprawl.data")
 in_rast      <- read_rast(in_file, bands = 20)
 tempraster   <- tempfile(fileext = ".tif")
 in_obj_zones <- raster::aggregate(in_rast,
                                     fact      = 4,
                                     filename  = tempraster,
                                     overwrite = TRUE)
 test <- aggregate_rast(in_rast,
                        in_obj_zones,
                        FUN     = mean)
}
\author{
Lorenzo Busetto, PhD (2017) email: \href{mailto:lbusett@gmail.com}{lbusett@gmail.com}
}
