% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate_rast.R
\name{aggregate_rast}
\alias{aggregate_rast}
\title{aggregate_rast}
\usage{
aggregate_rast(in_rast_values, in_obj_zones, FUN = mean,
  method = "fastdisk", to_file = FALSE, out_file = NULL, verbose = TRUE)
}
\arguments{
\item{in_rast_values}{PARAM_DESCRIPTION}

\item{in_obj_zones}{PARAM_DESCRIPTION}

\item{FUN}{PARAM_DESCRIPTION, Default: mean}

\item{method}{PARAM_DESCRIPTION, Default: 'fastdisk'}

\item{to_file}{PARAM_DESCRIPTION, Default: FALSE}

\item{out_file}{PARAM_DESCRIPTION, Default: NULL}

\item{verbose}{PARAM_DESCRIPTION, Default: TRUE}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
Function used to extract values of a single- or multi- band raster for pixels
corresponding to the features of a vector (Polygon, Point or Line)
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[sp]{proj4string}}
}
\author{
Lorenzo Busetto, phD (2017) \email{lbusett@gmail.com}
}
