% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_testshape.R
\name{build_testshape}
\alias{build_testshape}
\title{build_testshape}
\usage{
build_testshape(maxpolys, rmax = 10, allow_overlaps = FALSE,
  to_file = FALSE, ext = NULL, crs = NULL, seed = 100)
}
\arguments{
\item{maxpolys}{PARAM_DESCRIPTION}

\item{rmax}{PARAM_DESCRIPTION, Default: 10}

\item{allow_overlaps}{PARAM_DESCRIPTION, Default: FALSE}

\item{to_file}{PARAM_DESCRIPTION, Default: FALSE}

\item{ext}{PARAM_DESCRIPTION, Default: NULL}

\item{crs}{PARAM_DESCRIPTION, Default: NULL}

\item{seed}{PARAM_DESCRIPTION, Default: 100}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[tibble]{as_tibble}}
\code{\link[stats]{runif}}
}
