% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_testshape.R
\name{build_testshape}
\alias{build_testshape}
\title{build_testshape}
\usage{
build_testshape(maxpolys, allow_overlaps = FALSE, tofile = FALSE)
}
\arguments{
\item{maxpolys}{PARAM_DESCRIPTION}

\item{allow_overlaps}{PARAM_DESCRIPTION, Default: FALSE}

\item{to_file}{PARAM_DESCRIPTION, Default: FALSE}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
 p <- build_testshape(20)
}
}
\seealso{
\code{\link[dplyr]{filter}}
\code{\link[tibble]{as_tibble}}
}
