% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cast_rast.R
\name{cast_rast}
\alias{cast_rast}
\alias{cast_rast.default}
\alias{cast_rast.Raster}
\alias{cast_rast.character}
\title{automatic recasting between Raster object and raster files on disk}
\usage{
cast_rast(object, to)

\method{cast_rast}{default}(object, to)

\method{cast_rast}{Raster}(object, to)

\method{cast_rast}{character}(object, to)
}
\arguments{
\item{object}{either an \code{R} object of class \code{sf}, \code{sfc} or \code{Spatial}, or a character string
corresponding to a filename (with full path)}

\item{to}{\code{character} indicating to which type of object the input should be re-casted. It can
be "rastobject" (re-cast to \code{Raster}) or "rastfile" (recast to raster file).In the
second case, recasting is achieved by identifying the raster file associated with the
Raster object and returning its path. If no file is associated (i.e., the object is
in-memory), the Raster object is written to disk on a temporary file, which path is returned}
}
\value{
returns the same object, casted to the "class" specified by \code{to} (or the exact same
object in case \code{object} is already of "class" \code{to})
}
\description{
function to automatically re-cast between Raster object and raster files on disk
}
\details{
If \code{object} is a valid \code{R} spatial object, it is automatically converted to an object of
a different class if needed (e.g., from \code{sf} to \code{Spatial} and viceversa, or from \code{sf} to a
vector file through \code{sprawl::write_shape}). If it is a character string, the function checks if
it corresponds to a valid raster file and reads it to a \code{Raster} object through
\code{rasterr::stack()}
}
\examples{
 #EXAMPLE1
}
\author{
Lorenzo Busetto, phD (2017) \href{mailto:lbusett@gmail.com}{lbusett@gmail.com}
}
