% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/change_fileext.R
\name{change_fileext}
\alias{change_fileext}
\title{change file extension}
\usage{
change_fileext(in_file, new_ext, full_path = FALSE)
}
\arguments{
\item{in_file}{\code{character} full path of a file}

\item{new_ext}{\code{character} new file extension (e.g., ".png")}

\item{full_path}{\code{logical} if TRUE, a full path with updated extension is
returne, otherwise, only an updated basename is rerturned, Default: FALSE}
}
\value{
\code{character} basename/full_path of `in_file`` with updated extension
}
\description{
Helper function allowing to quckly change the extension of
a file name
}
\details{
DETAILS
}
\examples{
in_file = "home/lb/tmp/prova.tif"
# return an updated basename
change_fileext(in_file, ".png")

# return an updated full path
change_fileext(in_file, ".png", full_path = TRUE)

}
\author{
Lorenzo Busetto, phD (2017) \href{mailto:lbusett@gmail.com}{lbusett@gmail.com}
}
