% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_proj4string.R
\name{check_proj4string}
\alias{check_proj4string}
\alias{check_proj4string.default}
\alias{check_proj4string.character}
\alias{check_proj4string.CRS}
\title{Check the validity of the input proj4string}
\usage{
check_proj4string(proj4string, abort = FALSE)

\method{check_proj4string}{default}(proj4string, abort = FALSE)

\method{check_proj4string}{default}(proj4string, abort = FALSE)

\method{check_proj4string}{default}(proj4string)
}
\arguments{
\item{proj4string}{\code{character} or \code{\link{CRS}} corresponding to the
proj4string to be checked}

\item{abort}{\code{logical} if TRUE, the function aborts in case no proj4string or
invalid proj4string is found, Default: FALSE}
}
\value{
\code{character} proj4string of the object or file
}
\description{
helper function used to check that the input proj4string object
is a valid string or a CRS object.
}
\details{
DETAILS
}
\examples{
check_proj4string("+init:epsg:32632")

check_proj4string("example of invalid string")

check_proj4string(CRS("+init:epsg:32632"))

\dontrun{
 library(raster)
 in_rast <- system.file("extdata", "sprawl_EVItest.tif", package = "sprawl.data")
 in_rast <- raster::raster(in_rast)
 check_proj4string(in_rast@crs)
 }
}
\author{
Lorenzo Busetto, phD (2017) \href{mailto:lbusett@gmail.com}{lbusett@gmail.com}

Luigi Ranghetti, phD (2017) \href{mailto:ranghetti.l@irea.cnr.it}{ranghetti.l@irea.cnr.it}
}
