% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comp_zonal.R
\name{comp_zonal}
\alias{comp_zonal}
\title{comp_zonal}
\usage{
comp_zonal(in_rast, zone_object, selbands = NULL, mask_object = NULL,
  rastres = NULL, id_field = NULL, summ_data = TRUE, full_data = TRUE,
  comp_quant = FALSE, out_format = "dframe", small = TRUE, na.rm = TRUE,
  maxchunk = 5e+07, long = FALSE, addfeat = TRUE, addgeom = TRUE,
  keep_null = FALSE, verbose = TRUE, ncores = NULL)
}
\arguments{
\item{in_rast}{input raster. Can be wither:
\enumerate{
\item A file name corresponding to a valid single- or multi- band raster (e.g. /my_folder/myfile.tif)
All \code{gdal} raster formats are supported
\item An \code{R} \code{raster}, 'rasterStack\code{or}rasterBrick` object
}}

\item{zone_object}{input vector object. Can be either:
\enumerate{
\item A \code{file name} corresponding to a valid ESRI shapefile (e.g. /my_folder/myshape.shp)
\item An \code{R} \code{+sp} or \code{sf} object
}}

\item{selbands}{\code{2-element numeric array} defining starting and ending raster bands to be processed
(e.g., c(1, 10). Default: NULL, meaining that all bands will be processed}

\item{mask_object}{Not used for now - will allow to mask the raster before extraction of values, Default: NULL}

\item{rastres}{PARAM_DESCRIPTION, Default: NULL}

\item{id_field}{\code{character} (optional) field of the vector file to be used to identify the
zones from which data is to be extracted. MUST correspond to a univoc attribute (i.e., no replicated
values are allowed), If NULL (the default), zonal statistics are extracted on each row of the
shapefile, and a new column named \code{id_feat} is used on the output to distinguish the zones, Default: NULL}

\item{summ_data}{\code{logical} If TRUE, summary statistics of values belonging to each vector zone
are returned in \code{out$ts_summ~. Default statistics are average, median, standard deviation and variation coefficient. Quantiles of the distribution can be also returned if}comp_quant == TRUE`,  Default: TRUE}

\item{full_data}{\code{logical} If TRUE, values of all pixels belonging to each vector zone are returned
in `out$ts_full', Default: TRUE}

\item{comp_quant}{\code{logical} if TRUE, also quantiles of the distributions of values are computed for each zone
and returned in  \code{out$ts_summ}, Default: FALSE}

\item{out_format}{\code{character} either "dframe" or "xts" Define the output format, Default: 'dframe'}

\item{small}{\code{logical} if TRUE, values are returned also for small polygons not including any
pixel centroids, Default: TRUE}

\item{na.rm}{PARAM_DESCRIPTION, Default: TRUE}

\item{maxchunk}{PARAM_DESCRIPTION, Default: 5e+07}

\item{long}{PARAM_DESCRIPTION, Default: FALSE}

\item{addfeat}{PARAM_DESCRIPTION, Default: TRUE}

\item{addgeom}{PARAM_DESCRIPTION, Default: TRUE}

\item{keep_null}{PARAM_DESCRIPTION, Default: FALSE}

\item{verbose}{PARAM_DESCRIPTION, Default: TRUE}

\item{ncores}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
Function used to extract values of a single- or multi- band raster for pixels
corresponding to the features of a vector (Polygon, Point or Line)
}
\details{
DETAILS
}
\examples{

}
\author{
Lorenzo Busetto, phD (2014-2015) \email{lbusett@gmail.com}
}
