% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_rast_dtype.R
\name{convert_rastdtype}
\alias{convert_rastdtype}
\title{convert data type between \code{raster} and \code{gdal} conventions}
\usage{
convert_rastdtype(dtype_string, type)
}
\arguments{
\item{dtype_string}{\code{character} data type string (e.g., Int16, etc.)}

\item{type}{\code{character} - either "gdal" or "raster"}
}
\value{
\code{tibble} containing the representation of the data type both for gdal
and raster
}
\description{
FUNCTION_DESCRIPTION
}
\examples{
\dontrun{
 in_dtype <- "INT1U"
 convert_rastdtype(in_dtype, "raster")

 in_dtype <- "Float64"
 convert_rastdtype(in_dtype, "gdal")

 in_dtype <- "Float123"
 convert_rastdtype(in_dtype, "gdal")
}
}
\author{
Lorenzo Busetto, phD (2017) \href{mailto:lbusett@gmail.com}{lbusett@gmail.com}
}
