% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_fishnet.R
\name{create_fishnet}
\alias{create_fishnet}
\title{create_fishnet}
\usage{
create_fishnet(in_obj, cellsize, crop_layer = NULL, out_raster = NULL,
  out_shape = NULL, overwrite = TRUE, return_sp = TRUE, pypath = NULL)
}
\arguments{
\item{in_obj}{object of \code{sf} or \code{raster*} class from which the fishnet needs to be derived. The fishnet is
built from the corner of \code{in_obj}, with a resolution equal to \code{cellsize}}

\item{cellsize}{numeric Dimensions of the desired cells}

\item{crop_layer}{(optional) object of \code{Extent}. If not null, the fishnet extent will be cropped on
this extent, without "moving" the nodes of the grid. This is useful to crop a grid created on
the basis of a different raster coordinates on top of a different raster, Default: NULL}

\item{out_raster}{character (optional) name of the output raster file. If NULL the grid is NOT
saved in a raster file Default: NULL}

\item{out_shape}{character (optional) name of the output shape file. If NULL the grid is NOT
saved in a shapefile,  Default: NULL}

\item{overwrite}{\code{logical} If TRUE, existing files are overwritten, Default: FALSE}

\item{return_sp}{\code{logical}, If true, the grid is returned to the caller as a `SpatialGridDataFrame ,
Default: FALSE}

\item{pypath}{path to \code{gdal_polygonize.py}. If NULL, the path is searched on the user's machine,
Default: NULL}
}
\value{
if return_sp == TRUE, returns the grid as a SpatialGridDataFrame
}
\description{
funtion to create a raster and/or polygon fishnet of specified resolution over a
a specified spatial extent. If both in_ext and crop_ext are declared, the fishnet is created
according to in_ext (that is, the origin depends from in_ext), and then cropped on crop_ext (this
allows for example to create a fishnet based on a given raster/resolution, and then to automatically
crop it on a desired extgent without "moving" the corners of the cells ")
}
\details{
DETAILS
}
\examples{
\dontrun{
  # create a fishnet over an input raster file and save it as a shapefile
  library(raster)
  library(sprawl)

  file_in  <- system.file("extdata", "sprawl_EVItest.tif", package = "sprawl.data")
  inrast   <- raster::raster(file_in)
  cellsize <- raster::res(inrast)[1]

  out_file <- "/somedir/somefile.shp"
  fishnet  <- create_fishnet(inrast,
                             cellsize,
                             return_sp = TRUE
                             )

  plot(inrast)
  plot(fishnet, add = TRUE, col = "transparent")

 }
}
