% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_fishnet.R
\name{create_fishnet}
\alias{create_fishnet}
\title{Create a "fishnet" vector over the extent of a spatial object}
\usage{
create_fishnet(in_rast, pix_for_cell = 1, shape = "rect", cellsize = NULL,
  exact_csize = TRUE, to_file = FALSE, out_shape = NULL,
  overwrite = TRUE, crop_layer = NULL, verbose = TRUE)
}
\arguments{
\item{in_rast}{raster file or object of class \code{raster*} on which the fishnet
needs to be derived.}

\item{pix_for_cell}{\code{numeric(1/2)} (optional) 1/2 element numeric array
specifying how many pixels of the #' raster will be included in each
polygon of the fishnet (if only one element is provided, the #' same number
of pixel is aggregated in each direction). Ignored if \code{cellsize} is not null,
Default: 1}

\item{shape}{\code{character [\"rect\" \"hex\"}]}

\item{cellsize}{\code{numeric(1/2)} (optional) 1/2 element array specifying the
dimensions of the desired #' cells in the x and y directions (if only one
element is provided, the same cellsize is used in each direction),
Default: NULL}

\item{exact_csize}{\code{logical} If TRUE, the function strictly adheres to what
specified in \code{cellsize} or \code{pix_for_cell} (i.e., it doesn't alter the fishnet
cell to get a regualr fishnet over the raster). The last column/row of the
fisnhet have therefore a different area, but the other cells respect what
specified by the user (i.e., having a regular 2x2 km grid).}

\item{to_file}{\code{logical} PARAM_DESCRIPTION, Default: FALSE.}

\item{out_shape}{\code{logical} PARAM_DESCRIPTION, Default: FALSE.}

\item{overwrite}{\code{logical} PARAM_DESCRIPTION, Default: TRUE.}

\item{crop_layer}{\code{logical} object of class \code{Extent}. If not null, the
fishnet is cropped on this extent, without "moving" the nodes of the grid.
This is useful to crop a grid created on the basis of a different raster
coordinates on top of a different raster, Default: FALSE.}

\item{verbose}{\code{logical} If FALSE, processing messages are suppressed,
Default: TRUE.}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
Function to create a polygon fishnet of specified resolution
over a raster dataset. By default, the fishnet is built so that
each cell corresponds to a raster cell. If the \code{pix_for_cell} argument is
set, then the polygons of the fishnet are built so to include the specified
number of cells in each direction. If the \code{cellsize} option is set, a fishnet
of the specified resolution is created over the extent of the raster)
}
\details{
DETAILS
}
\examples{
\dontrun{
  # create a fishnet over an input raster file and save it as a shapefile
  library(raster)
  library(sprawl)
  library(rasterVis)

  in_rast  <- build_testraster(20,40,1)
  fishnet  <- create_fishnet(in_rast)

  plot_rast(in_rast, in_poly = fishnet)

  fishnet  <- create_fishnet(in_rast, pix_for_cell = c(4,2))
  plot_rast(in_rast, in_poly = fishnet)

  fishnet  <- create_fishnet(in_rast, cellsize = c(25,25))
  plot_rast(in_rast, in_poly = fishnet)

  # plotting with `exact_csize` creates a grid that covers the extent with
  # regular cells, by adapting the cellsize (see `sf::st_make_grid`)
  fishnet  <- create_fishnet(in_rast, cellsize = c(25,25),
                                  exact_csize = F)
  plot_rast(in_rast, in_poly = fishnet)

  # using shape = "hex" gives and hexagonal grid instead
  fishnet  <- create_fishnet(in_rast, cellsize = c(25,25),
                                  exact_csize = T, shape = "hex")
  plot_rast(in_rast, in_poly = fishnet)
 }
}
\author{
Lorenzo Busetto, PhD (2017) email: \href{mailto:lbusett@gmail.com}{lbusett@gmail.com}
}
