% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_virtrast.R
\name{create_virtrast}
\alias{create_virtrast}
\title{create a GDAL vrt file starting from a \code{Raster} object}
\usage{
create_virtrast(object, out_vrt_file, out_extent = NULL, verbose = TRUE)
}
\arguments{
\item{object}{PARAM_DESCRIPTION}

\item{out_vrt_file}{PARAM_DESCRIPTION}

\item{out_extent}{PARAM_DESCRIPTION, Default: NULL}

\item{verbose}{PARAM_DESCRIPTION, Default: NULL}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
create a GDAL vrt file starting from a \code{Raster} object
}
\details{
DETAILS
}
\examples{

 #EXAMPLE1

}
\author{
Lorenzo Busetto, phD (2017) \href{mailto:lbusett@gmail.com}{lbusett@gmail.com}
}
