% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crop_rast.R
\name{crop_rast}
\alias{crop_rast}
\title{crop a raster file/rast_file on a given extent}
\usage{
crop_rast(rast_object, ext_object, mask = FALSE, pad = 1,
  out_type = "rastobject", out_file = NULL, out_dtype = NULL,
  compress = "None", verbose = TRUE)
}
\arguments{
\item{rast_object}{either an \code{R} object of class \code{Raster}, or a character string
corresponding to a raster filename (with full path)}

\item{ext_object}{either an object of class \code{sprawlext}, or an \code{R} object or
filename from which an object of class \code{sprawlext} can be obtained (see
\code{sprawl::get_extent()})}

\item{mask}{\code{logical} if TRUE, and \code{ext_object} is a polygon shapefile or
\code{R} object, the output is also masked on the boundaries of \code{ext_object}}

\item{pad}{\code{numeric} extent of a "padding area" to be kept with respect to
\code{crop_extent} (in number of pixels). Useful to be sure not to crop "too much"
for example when cropping before extracting features suche as in \code{extract_rast},
Default: 1}

\item{out_type}{\code{character} indicates the type of object to be returned:
\enumerate{
\item "rastobject" return a \code{raster} rast_file accessing the saved cropped file
\item "rastfile" return the filename of the GTiff file corresponding to the
cropped GTiff file. If \code{out_file == NULL}, this corresponds to a file
saved in \code{R} temporary folder.
\item "vrtfile" return the filename of the vrt file built for cropping
the input raster (\strong{no saving to disk is performed})
, Default: "rastobject"
}}

\item{out_file}{\code{character} filename to be used to save the cropped
raster.}

\item{out_dtype}{\code{character} data type of the output masked files, according
to gdal specifications for GTiff files ("Byte", "UInt16", "Int16", "UInt32",
"Int32", "Float32", "Float64", "CInt16", "CInt32", "CFloat32" and "CFloat64").
If NULL, the data type is retrieved from the input, Default: NULL}

\item{compress}{\code{character} compression option to be used to saved the cropped
raster ("None", "PACKBITS", "LZW", "DEFLATE), Default: "None"}

\item{verbose}{\code{logical} if FALSE, suppress processing messages, Default: TRUE}
}
\value{
either a \code{Raster} object containing the cropped raster, or a gdal vrt
or GTiff filename corresponding to it, depending on \code{out_type}.
}
\description{
function to crop a raster file/rast_file on a given extent.
The cropping extent can be:
\enumerate{
\item directly passed (as a \code{sprawlext} object or numeric array);
\item derived from a different object from which a \code{sprawlext} object object
can be derived (see details)
}
}
\examples{

in_file <- system.file("extdata/OLI_test", "oli_multi_1000.tif",
                      package = "sprawl.data")
in_rast <- read_rast(in_file)

in_vect     <- create_fishnet(in_rast, pix_for_cell = 60,
                              verbose = FALSE)[55:60,]

plot_rast(in_rast[[1]], in_poly = in_vect)

out_cropped <- crop_rast(in_rast, in_vect)

get_extent(in_rast)
get_extent(out_cropped)

plot_rast(out_cropped[[1]])

}
\author{
Lorenzo Busetto, phD (2017) \href{mailto:lbusett@gmail.com}{lbusett@gmail.com}
}
