% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crop_vect.R
\name{crop_vect}
\alias{crop_vect}
\title{crop a vector on the extent of a different spatial file}
\usage{
crop_vect(in_vect, in_obj, to_file = FALSE, as_sp = FALSE)
}
\arguments{
\item{in_vect}{PARAM_DESCRIPTION}

\item{in_obj}{PARAM_DESCRIPTION}

\item{to_file}{PARAM_DESCRIPTION, Default: FALSE}

\item{as_sp}{PARAM_DESCRIPTION, Default: FALSE}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
Function allowing to crop an input vector object or shapefile on the extent of a
different spatial object. in_vect must be an *sp, *sf object or vector file, while in_obj should
be a valid spatial object (sp, sf, vector file, raster or rasterfile)
}
\details{
DETAILS
}
\examples{
\dontrun{
 #EXAMPLE1
 }
}
\seealso{
\code{\link[sp]{proj4string}}
}
