% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crop_vect.R
\name{crop_vect}
\alias{crop_vect}
\title{crop a vector on the extent of a different spatial file}
\usage{
crop_vect(in_vect, ext_obj, out_file = NULL, verbose = TRUE)
}
\arguments{
\item{in_vect}{Object of class \code{sfc_POLYGON} or \code{sf_POLYGON}, or coercible to
it using (sprawl::cast_vect)}

\item{ext_obj}{Object of class \code{sprawl_ext}, or from which an extent can be
retrieved using \code{spralw::get_extent} (i.e., any \code{R} spatial object, or
valid filenames of raster or vector files)}

\item{out_file}{PARAM_DESCRIPTION, Default: NULL}

\item{verbose}{PARAM_DESCRIPTION, Default: FALSE}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
Function allowing to crop an input vector object or shapefile on the extent of a
different spatial object. in_vect must be an *sp, *sf object or vector file, while ext_obj should
be a valid spatial object (sp, sf, vector file, raster or raster file name)
}
\details{
DETAILS
}
\examples{
\dontrun{
 library(raster)
 library(sf)
 in_polys <- read_vect(system.file("extdata/shapes", "lc_polys.shp",
                                         package = "sprawl.data"))
 in_rast <- raster::raster(system.file("extdata/MODIS_test",
                                  "EVIts_test.tif", package = "sprawl.data"))
 in_polys_cropped <- crop_vect(in_polys, in_rast)

 # Retrieve a polygon representing in_rast extent and reproject it
 rast_extent <- as(get_extent(in_rast), "sfc_POLYGON") \%>\%
    sf::st_transform(get_proj4string(in_polys))
 par(mfrow=c(1,2))
 plot(in_polys[1])
 plot(rast_extent, add = T)
 plot(in_polys_cropped[1])
 plot(rast_extent, add = T)
  par(mfrow=c(1,2))
 }
}
\seealso{
\code{\link[sp]{proj4string}}
}
