% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cz_polygons.R
\name{cz_polygons}
\alias{cz_polygons}
\title{cz_polygons}
\usage{
cz_polygons(zone_object, in_rast, seldates, selbands, n_selbands, date_check,
  cz_opts)
}
\arguments{
\item{zone_object}{PARAM_DESCRIPTION}

\item{in_rast}{PARAM_DESCRIPTION}

\item{seldates}{PARAM_DESCRIPTION}

\item{selbands}{PARAM_DESCRIPTION}

\item{n_selbands}{PARAM_DESCRIPTION}

\item{date_check}{PARAM_DESCRIPTION}

\item{cz_opts}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[doSNOW]{registerDoSNOW}}
\code{\link[foreach]{foreach}}
\code{\link[gdalUtils]{gdalwarp}}
\code{\link[sp]{proj4string}}
\code{\link[velox]{velox}}
\code{\link[utils]{setTxtProgressBar}},\code{\link[utils]{txtProgressBar}}
}
