% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_doydates.R
\name{datetodoy}
\alias{datetodoy}
\title{Convert dates to DOYs}
\usage{
datetodoy(dates = dates)
}
\arguments{
\item{dates}{array of dates to be converted. Class must be \code{Date} or character
parsable to \code{Date}
using \code{\link{as.Date}} (format "yyyy-mm-dd").}
}
\description{
Converts an array of \code{Date} (or coercible to \code{Date}) values to
DOY (Day Of the Year) values.
}
\author{
Lorenzo Busetto, PhD (2017) email: \href{mailto:lbusett@gmail.com}{lbusett@gmail.com}
}
