% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datesconversion.R
\name{datetodoy}
\alias{datetodoy}
\title{datetodoy}
\usage{
datetodoy(dates = dates)
}
\arguments{
\item{dates}{array of dates to be converted. Class must be \code{Date} or character parsable to \code{Date}
using \code{\link{as.Date}} (format "yyyy-mm-dd")}
}
\value{
array of DOYs resulting from the conversion
}
\description{
Converts an array of \code{Date} (or cohercible to \code{Date})values to DOY (Day Of the Year)
values (simple wrapper of 'strftime(date, format = '%j')`)
}
