% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dissolve_vect.R
\name{dissolve_vect}
\alias{dissolve_vect}
\title{dissolve attributes of a vector and aggregate its table}
\usage{
dissolve_vect(in_vect, dissolve_var, var_as_NA = FALSE, out_file = NULL,
  out_type = "vectobject", overwrite = FALSE, verbose = TRUE)
}
\arguments{
\item{in_vect}{Either a \code{R} vector object (\code{sf} or \code{sp}), or the path to
a valid vector file}

\item{dissolve_var}{\code{character} column of the attribute table of
\code{in_vect} on to which the dissolve has to be performed.}

\item{var_as_NA}{\code{logical} TO BE DESCRIBED - see examples for now!!!,
Default: FALSE}

\item{out_file}{\code{character} output file name}

\item{out_type}{\code{character ["vectfile" | "vectobject"]} If "vectfile", and \code{out_file}
is not NULL, the function returns the name of the saved shapefile. Otherwise,
it returns the reprojected vector object, with the format specified by
out_class,  Default: 'vectobject'}

\item{overwrite}{\code{logical} if TRUE, output file will be overwritten if existing,
Default: FALSE}

\item{verbose}{\code{logical} if TRUE, provide messages on processing, Default: TRUE}
}
\value{
an \code{*sf} object. The object is also saved to disk as a shapefile if
\code{out_file} is provided
}
\description{
Perform a "dissolve" (a.k.a. aggregation) operation on a vector
object according to a grouping variable, also aggregate values of the
attribute table for all other columns over the same grouping variable.

Aggregation of factor or character columns is done as follows:
- If all polygons that are aggregated to a common dissolved polygon had
the same value for the column, then the value is preserved;
- If polygons had different values, the column takes the value "variabl"
or NA (depending on \code{var_as_NA} )

Aggregation of factor or character columns is done as follows:\preformatted{- If the column represent a "quantity" (e.g., Population), the value for each
  dissolved is computed as the sum of the values of the original
  polygons belonging to it;
}

TODO
- If the column represents a "density" (e.g., ton ha-1), the value for each
dissolved is computed as the average of the values of the original
polygons belonging to it, weighted by their area;  STILL TO BE IMPLEMENTED !
CURRENTLTY ALL COLUMNS ARE COMPUTED SIMPLY aS THE SUM !!!!
}
\examples{

 library(sprawl.data)
 library(maps)

 in_vect    <- get(load(system.file("extdata/shapes","poly_lomb.RData",
    package = "sprawl.data")))
 plot_vect(in_vect, fill_var = "NAME_2")
 plot_vect(in_vect, fill_var = "population", palette_name = "RdYlGn")

 # Dissolve the vector to provinces
 diss <- dissolve_vect(in_vect, "NAME_2")
 diss

 plot_vect(diss, fill_var = "NAME_2")
 provinces <- get_boundaries("ITA", 2)

 plot_vect(diss, fill_var = "population", palette_name = "RdYlBu",
         leg_position = "bottom")


}
\author{
Lorenzo Busetto, phD (2017) \href{mailto:lbusett@gmail.com}{lbusett@gmail.com}
}
