% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datesconversion.R
\name{doytodate}
\alias{doytodate}
\title{doytodate}
\usage{
doytodate(doys, year, verbose = TRUE)
}
\arguments{
\item{doys}{numeric array DOYs to be converted. Range must be between 0 and 366}

\item{year}{numeric year to be used as basis.}

\item{verbose}{'logical' if TRUE, print messages, default = TRUE}
}
\value{
\code{Date} array resulting from the conversion
}
\description{
Converts a doy (or array of doys) to date format.
Year of origni has to be specified
}
