% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_doydates.R
\name{doytodate}
\alias{doytodate}
\title{convert DOYs to dates}
\usage{
doytodate(doys, year, verbose = TRUE)
}
\arguments{
\item{doys}{'numeric array' DOYs to be converted.}

\item{year}{'numeric year to be used as basis.}

\item{verbose}{'logical' if TRUE, print messages, default = TRUE}
}
\value{
\code{Date} array resulting from the conversion
}
\description{
Converts a doy (or array of doys) to date format. Year of origin has to be specified, to allow proper handling of leap years
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\author{
Lorenzo Busetto, PhD (2017) email: \href{mailto:lbusett@gmail.com}{lbusett@gmail.com}
}
