% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/er_points.R
\name{er_points}
\alias{er_points}
\title{helper for extract_rast for points extraction}
\usage{
er_points(in_vect_points, in_rast, n_selbands, selbands, seldates, id_field,
  long, date_check, verbose, addfeat, addgeom, keep_null)
}
\arguments{
\item{in_vect_points}{\code{sf} object on which features the zonal statistics has to be computed}

\item{in_rast}{single- or multi-band \code{rasterStack} object containing values to be extracted
and summarized}

\item{n_selbands}{PARAM_DESCRIPTION}

\item{selbands}{PARAM_DESCRIPTION}

\item{seldates}{PARAM_DESCRIPTION}

\item{id_field}{PARAM_DESCRIPTION}

\item{date_check}{PARAM_DESCRIPTION}

\item{verbose}{PARAM_DESCRIPTION}

\item{addfeat}{PARAM_DESCRIPTION}

\item{addgeom}{PARAM_DESCRIPTION}

\item{keep_null}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\author{
Lorenzo Busetto, phD (2017) \href{mailto:lbusett@gmail.com}{lbusett@gmail.com}
}
