% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/er_polygons.R
\name{er_polygons}
\alias{er_polygons}
\title{extract raster data on polygons (helper for extract_rast)}
\usage{
er_polygons(in_vect_crop, in_vect, in_rast, na.value, seldates, selbands,
  n_selbands, date_check, er_opts, outside_feat, verb_foreach = FALSE)
}
\arguments{
\item{in_vect_crop}{PARAM_DESCRIPTION}

\item{in_vect}{PARAM_DESCRIPTION}

\item{in_rast}{PARAM_DESCRIPTION}

\item{seldates}{PARAM_DESCRIPTION}

\item{selbands}{PARAM_DESCRIPTION}

\item{n_selbands}{PARAM_DESCRIPTION}

\item{date_check}{PARAM_DESCRIPTION}

\item{er_opts}{PARAM_DESCRIPTION}

\item{outside_feat}{PARAM_DESCRIPTION}

\item{verb_foreach}{\code{logical} if TRUE, verbose output is sent out from within the foreach cycle,
Default: FALSE}

\item{na.vlaue}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
 #EXAMPLE1
 }
}
