% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_gdal.R
\name{find_gdal}
\alias{find_gdal}
\title{find gdal main folder}
\usage{
find_gdal()
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{

 find_gdal()

}
\author{
Lorenzo Busetto, phD (2017) \href{mailto:lbusett@gmail.com}{lbusett@gmail.com}
}
