% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_boundaries.R
\name{get_boundaries}
\alias{get_boundaries}
\title{Download vector of administrative boundaries from GADM}
\usage{
get_boundaries(iso, level = 1, path = tempdir())
}
\arguments{
\item{iso}{\code{character} iso name or 2/3 digits code of the country}

\item{level}{\code{numeric} level of administrative units returned, Default: 0
(Country Level)}

\item{path}{\code{character} folder where the "RData" file containing polygon
boundaries should be saved, Default: \code{tempdir()}}
}
\value{
\code{sf} object containing the desired data
}
\description{
Function to download administrative boundaries data from the gadm
archive, starting from a country name or ISO code, and the level of desired
administrative units. The function is a simple wrapper to the
\code{raster::getData} function, with a-priori checks on country name or ISO code
and automatic recasting to \code{sf} format.
}
\examples{
\dontrun{
 library(sprawl)
 ita_boundaries <- get_boundaries("Italy", 0)
 ita_region_boundaries <- get_boundaries("ITA", 1)
 plot(ita_region_boundaries[1])
 }
}
\seealso{
\code{\link[raster]{getData}}
}
