% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_extent.R
\name{get_extent}
\alias{get_extent}
\alias{get_extent.default}
\alias{get_extent.sprawlext}
\alias{get_extent.Extent}
\alias{get_extent.matrix}
\alias{get_extent.character}
\alias{get_extent.Raster}
\alias{get_extent.sf}
\alias{get_extent.sfc}
\alias{get_extent.Spatial}
\title{return the extent of a spatial object or file (with projection)}
\usage{
get_extent(in_object, proj4string = NULL, abort = FALSE)

\method{get_extent}{default}(in_object, proj4string = NULL,
  abort = FALSE)

\method{get_extent}{sprawlext}(in_object, proj4string = NULL,
  abort = FALSE)

\method{get_extent}{Extent}(in_object, proj4string, abort = FALSE)

\method{get_extent}{matrix}(in_object, proj4string, abort = FALSE)

\method{get_extent}{character}(in_object, proj4string = NULL,
  abort = FALSE)

\method{get_extent}{Raster}(in_object, proj4string = NULL,
  abort = FALSE)

\method{get_extent}{sf}(in_object, proj4string = NULL, abort = FALSE)

\method{get_extent}{sfc}(in_object, proj4string = NULL, abort = FALSE)

\method{get_extent}{Spatial}(in_object, proj4string = NULL,
  abort = FALSE)

\method{get_extent}{Spatial}(in_object, proj4string = NULL,
  abort = FALSE)
}
\arguments{
\item{in_object}{\code{character} corresponding to the name of a R object, or a
filename (full path)}

\item{proj4string}{(optional) \code{character} proj4string representing the projection of
the input extent. It is needed only if \code{in_object} does not include a
projection (like \code{\link{extent}} or \code{\link{bbox}}).}

\item{abort}{\code{logical} if TRUE, the function aborts in case valid a extent can
not be retrieved, otherwise, it issues a Warning and returns \code{NA},
Default: FALSE}
}
\value{
object of class \code{sprawlext}
}
\description{
helper function used to retrieve the extent of a spatial object
or file, and the associated proj4string
}
\details{
return an object of class \code{sprawlext} with two slots:
\itemize{
\item $extent: \code{numeric (4)} extent of the object (xmin, ymin, xmax, ymax)
\item $proj4string: \code{character} proj4string of the object
(if an extent can not be retrieved from \code{in_object})
}
}
\examples{

library(raster)

in_rast <- system.file("extdata/MODIS_test", "EVIts_test.tif",
  package = "sprawl.data")
get_extent(in_rast)

in_rast <- raster::raster(in_rast)
get_extent(in_rast)

in_vect <- system.file("extdata/shapes","lc_polys.shp",
  package = "sprawl.data")
get_extent(in_vect)

in_vect <- read_vect(in_vect)
get_extent(in_vect)

}
\author{
Lorenzo Busetto, phD (2017) \href{mailto:lbusett@gmail.com}{lbusett@gmail.com}

Luigi Ranghetti, phD (2017) \href{mailto:ranghetti.l@irea.cnr.it}{ranghetti.l@irea.cnr.it}
}
