% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_proj4string.R
\name{get_proj4string}
\alias{get_proj4string}
\alias{get_proj4string.default}
\alias{get_proj4string.character}
\alias{get_proj4string.Raster}
\alias{get_proj4string.sf}
\alias{get_proj4string.sfc}
\alias{get_proj4string.Spatial}
\title{return the proj4string of a spatial object or file}
\usage{
get_proj4string(object, abort = FALSE)

\method{get_proj4string}{default}(object, abort = FALSE)

\method{get_proj4string}{character}(object, abort = FALSE)

\method{get_proj4string}{Raster}(object, abort = FALSE)

\method{get_proj4string}{sf}(object, abort = FALSE)

\method{get_proj4string}{sfc}(object, abort = FALSE)

\method{get_proj4string}{Spatial}(object, abort = FALSE)
}
\arguments{
\item{object}{\code{character} corresponding to the name of a R object, or a filename
(full path)}

\item{abort}{\code{logical} if TRUE, the function aborts in case no proj4string or
invalid proj4string is found, Default: FALSE}
}
\value{
\code{character} proj4string of the object or file
}
\description{
helper function used to extract the proj4string of "R" spatial objects
or of raster or vector files
}
\details{
DETAILS
}
\examples{
\dontrun{
library(raster)

in_rast <- system.file("extdata/MODIS_test", "EVIts_test.tif", package = "sprawl.data")
get_proj4string(in_rast)

in_rast <- raster::raster(in_rast)
get_proj4string(in_rast)

in_vect <- system.file("extdata/shapes","lc_polys.shp", package = "sprawl.data")
get_proj4string(in_vect)

in_vect <- read_vect(in_vect)
get_proj4string(in_vect)
}

}
\author{
Lorenzo Busetto, phD (2017) \href{mailto:lbusett@gmail.com}{lbusett@gmail.com}

Luigi Ranghetti, phD (2017) \href{mailto:ranghetti.l@irea.cnr.it}{ranghetti.l@irea.cnr.it}
}
