% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_projstring.R
\name{get_projstring}
\alias{get_projstring}
\alias{get_projstring.character}
\alias{get_projstring.Raster}
\alias{get_projstring.sf}
\alias{get_projstring.sfc}
\alias{get_projstring.Spatial}
\title{return the proj4string of a spatial object or file}
\usage{
get_projstring(object)

\method{get_projstring}{character}(object)

\method{get_projstring}{Raster}(object)

\method{get_projstring}{sf}(object)

\method{get_projstring}{sfc}(object)

\method{get_projstring}{Spatial}(object)
}
\arguments{
\item{object}{\code{character} corresponding to the name of a R object, or a filename
(full path)}
}
\value{
\code{character} proj4string of the object or file
}
\description{
helper function used to extract the proj4string of "R" spatial objects
or of raster or vector files
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 library(raster)

 in_rast <- system.file("extdata", "sprawl_EVItest.tif", package = "sprawl.data")
 get_projstring(in_rast)

 in_rast <- raster::raster(in_rast)
 get_projstring(in_rast)

 in_vect <- system.file("extdata","lc_polys.shp", package = "sprawl.data")
 get_projstring(in_vect)

 in_vect <- read_vect(in_vect)
 get_projstring(in_vect)

 }
}
}
\author{
Lorenzo Busetto, phD (2017) \href{mailto:lbusett@gmail.com}{lbusett@gmail.com}
}
