% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_rastinfo.R
\name{get_rastinfo}
\alias{get_rastinfo}
\title{Retrieve useful info from a raster object or file}
\usage{
get_rastinfo(object)
}
\arguments{
\item{object}{either the name of a \code{Raster} object or a valid raster filename}
}
\value{
\code{list} containing the following object:
\itemize{
\item nbands: \code{numeric} Number of bands of the object/file;
\item indbands: \code{numeric} If the object is associated with a file on disk, bands of the
file to which each layer of the object correspond (see examples);
\item ncols: \code{numeric} Number of columns;
\item nrows: \code{numeric} Number of rows;
\item ncells: \code{numeric} Number of cells;
\item bnames: \code{character} Band names;
\item fnames: \code{character} If the object is associated with one or more files on disk,
filename(s) corresponding to the different bands;
\item Z: \code{ANY} if the object has a "z" attribute set using raster::setZ(), values of the
z attribute;
\item dtype: \code{character} datatype of the input raster (\code{raster} conventions)
\item proj4string: \code{character} proj4string of the object/file
\item units: \code{character} distance units of the projections
}
}
\description{
Function used to facilitate retrieval useful info from a raster object or
file.
}
\examples{
\dontrun{
 in_rast <- system.file("extdata/OLI_test", "oli_multi_1000.tif",
  package = "sprawl.data")
 get_rastinfo(in_rast)
 in_rast <- system.file("extdata/OLI_test", "oli_multi_1000_b1.tif",
  package = "sprawl.data")
 get_rastinfo(in_rast)
 }
}
\author{
Lorenzo Busetto, phD (2017) \href{mailto:lbusett@gmail.com}{lbusett@gmail.com}
}
