% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_rastinfo.R
\name{get_rastinfo}
\alias{get_rastinfo}
\title{Retrieve useful info from a raster object or file}
\usage{
get_rastinfo(in_rast, stats = FALSE, quantiles = FALSE, hist = FALSE,
  verbose = TRUE)
}
\arguments{
\item{in_rast}{a \code{Raster} object or the filename of a valid raster}

\item{stats}{\code{logical} If TRUE, retrieve also statistics (min, max, mean, etc.)
for each band, Default: TRUE}

\item{quantiles}{\code{logical} If TRUE, retrieve also the quantiles, for each band,
Default; FALSE}

\item{hist}{\code{logical} If TRUE, retrieve also the frequency histogram, for
each band, Default: FALSE}

\item{verbose}{if FALSE suppress messages}
}
\value{
\code{list} containing the following object:
\itemize{
\item \strong{nbands}: \code{numeric} Number of bands of the object/file;
\item \strong{indbands}: \code{numeric} If the object is associated with a file on disk, bands of the
file to which each layer of the object correspond (see examples);
\item \strong{ncols}: \code{numeric} Number of columns;
\item \strong{nrows}: \code{numeric} Number of rows;
\item \strong{ncells}: \code{numeric} Number of cells;
\item \strong{bnames}: \code{character} Band names;
\item \strong{fnames}: \code{character} If the object is associated with one or more files on disk,
filename(s) corresponding to the different bands;
\item \strong{Z}: \code{ANY} if the object has a "z" attribute set using raster::setZ(), values of the
z attribute;
\item \strong{dtype}: \code{character} datatype of the input raster (\code{raster} conventions)
\item \strong{proj4string}: \code{character} proj4string of the object/file
\item \strong{units}: \code{character} distance units of the projections
}

According to settings on \code{stats}, \code{quantiles} and \code{hist}, the following
info are also reported:
\itemize{
\item \strong{stats}: \code{data.frame} containing min, max, average and standard deviation
for each band (see examples);
\item \strong{quants}: \code{data.frame} containing the quantiles of the distribution
of raster values, for each band (100 value - 0.01 interval) (NULL is returned
if \code{quants} == FALSE (the default));
\item \strong{hists}: \code{data.frame} containing information about the distribution of
raster values for each band. the data frame includes the limits of each
bin, the count of the number of pixels included in it, the corresponding
frequency and the cumulated frequency at each bin  (NULL is returned
if \code{hists} == FALSE (the default));
}
}
\description{
Function used to facilitate retrieval useful info from a raster object or
file.
}
\examples{

 in_rast <- system.file("extdata/OLI_test", "oli_multi_1000.tif",
  package = "sprawl.data")
 get_rastinfo(in_rast)
 in_rast <- system.file("extdata/OLI_test", "oli_multi_1000_b1.tif",
  package = "sprawl.data")
 get_rastinfo(in_rast)

}
\author{
Lorenzo Busetto, phD (2017) \href{mailto:lbusett@gmail.com}{lbusett@gmail.com}
}
