% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_rastype.R
\name{get_rastype}
\alias{get_rastype}
\title{Check in the input is a \code{Raster} object or raster file}
\usage{
get_rastype(in_object, abort = TRUE)
}
\arguments{
\item{in_object}{\code{R} object or path to a file}

\item{abort}{If TRUE, and \code{in_object} is neither a raster object or
filename, send an error message and abort, Default: TRUE}
}
\value{
\code{character} equal to "rastfile" (if \code{in_object} is a raster file),
\code{rastobject} (if \code{in_object} is a \code{R} raster object) or "none" if it is
neither (unless \code{abort} == TRUE)
}
\description{
Check if a \code{R} object or a filename correspond to a valid \code{Raster}
object, to a raster file or none of the above. Useful to detect which kind
of input is passed to a function and abort / do something in the case of
"wrong" input.
}
\details{
DETAILS
}
\examples{

#EXAMPLE1

}
\author{
Lorenzo Busetto, phD (2017) \href{mailto:lbusett@gmail.com}{lbusett@gmail.com}
}
