% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_spatype.R
\name{get_spatype}
\alias{get_spatype}
\title{check the "spatial type" of an object or file}
\usage{
get_spatype(in_object, abort = TRUE)
}
\arguments{
\item{in_object}{either a \code{R} object or a \code{character} string pointing to a
vector or raster layer}

\item{abort}{\code{logical} if TRUE the function aborts if \code{object} is not
recognized as an R spatial file or valid vector or raster file; if FALSE,
a warning is shown and `NA`` is returned.}
}
\value{
character ("\emph{spobject}" | "\emph{sfobject}" | "\emph{rastobject} | "
\emph{vectfile}" | \emph{rastfile} " | \emph{sprawlext}), or \code{NA} if the input does not
belong to any spatial category and abort == FALSE
}
\description{
accessory function to check if an object passed to the function
corresponds to a \code{*Spatial} Object, a \code{sf} object, a R \code{raster} object, a
file corresponding to a vector, or a file corresponding to a raster.
NOTE: to check only for vector or raster types, the associated functions
\code{get_vectype} and \code{get_rastype} can be used, with the same syntax.
}
\examples{

library(sprawl.data)
# input is a raster file
in_rast <- system.file("extdata/MODIS_test", "EVIts_test.tif",
                        package = "sprawl.data")
get_spatype(in_rast)

# input is a raster object
obj <- read_rast(in_rast)
get_spatype(in_rast)

# input is a shapefile
in_vect <- system.file("extdata/shapes","lc_polys.shp",
           package = "sprawl.data")
get_spatype(in_vect)

# input is a `sp` object
obj <- read_vect(in_vect, as_sp = TRUE)
get_spatype(obj)

# input is a `sf` object
obj <- read_vect(in_vect)
get_spatype(obj)

# input is a `sprawlext` object
obj <- get_extent(in_vect)
get_spatype(obj)
}
\author{
Lorenzo Busetto, phD (2017) \href{mailto:lbusett@gmail.com}{lbusett@gmail.com}

Luigi Ranghetti, phD (2017) \href{mailto:ranghetti.l@irea.cnr.it}{ranghetti.l@irea.cnr.it}
}
