% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_folder.R
\name{make_folder}
\alias{make_folder}
\title{FUNCTION_TITLE}
\usage{
make_folder(path, type = "dirname", verbose = TRUE)
}
\arguments{
\item{path}{PARAM_DESCRIPTION}

\item{type}{PARAM_DESCRIPTION, Default: 'dirname'}

\item{verbose}{PARAM_DESCRIPTION, Default: 'FALSE'}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
path <- file.path(tempdir(), "testdir")
dir.exists(path)
make_folder(path)
dir.exists(dirname(path))

path <- file.path(tempdir(), "test2", "testfile.xyz")
dir.exists(dirname(path))
make_folder(path, type = "filename")
dir.exists(dirname(path))

}
}
\author{
Lorenzo Busetto, phD (2017) \href{mailto:lbusett@gmail.com}{lbusett@gmail.com}
}
