% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mask_rast.R
\name{mask_rast}
\alias{mask_rast}
\title{mask_rast}
\usage{
mask_rast(in_rast, mask_shape, buffer = NULL, verbose = TRUE)
}
\arguments{
\item{in_rast}{PARAM_DESCRIPTION}

\item{mask_shape}{PARAM_DESCRIPTION}

\item{buffer}{PARAM_DESCRIPTION, Default: NULL}

\item{verbose}{PARAM_DESCRIPTION, Default: TRUE}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
library(raster)
in_polys <- read_shape(system.file("extdata","lc_polys.shp", package = "sprawl"), stringsAsFactors = T)
in_rast  <- raster::stack(system.file("extdata", "testrast.tif", package = "sprawl"))[[1]]
in_polys <- sf::st_transform(in_polys, proj4string(in_rast))
masked   <- mask_rast(in_rast, in_polys, verbose = FALSE)
par(mfrow=c(1,2))
plot(in_rast)
plot(masked)
}
}
\seealso{
\code{\link[dplyr]{case_when}}
\code{\link[gdalUtils]{gdal_rasterize}}
\code{\link[raster]{raster}},\code{\link[raster]{origin}}
}
