% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_rast_gg.R
\name{plot_rast_gg}
\alias{plot_rast_gg}
\title{FUNCTION_TITLE}
\usage{
plot_rast_gg(in_rast, xlims = NULL, ylims = NULL, basemap = NULL,
  zoomin = -1, scalebar = TRUE, scalebar_dist = NULL,
  na.color = "transparent", palette_type = "gradient", palette = NULL,
  no_labels = TRUE, title = NULL, subtitle = NULL)
}
\arguments{
\item{in_rast}{PARAM_DESCRIPTION}

\item{xlims}{PARAM_DESCRIPTION, Default: NULL}

\item{ylims}{PARAM_DESCRIPTION, Default: NULL}

\item{basemap}{PARAM_DESCRIPTION, Default: NULL}

\item{zoomin}{PARAM_DESCRIPTION, Default: -1}

\item{scalebar}{PARAM_DESCRIPTION, Default: TRUE}

\item{scalebar_dist}{PARAM_DESCRIPTION, Default: NULL}

\item{na.color}{PARAM_DESCRIPTION, Default: 'transparent'}

\item{palette_type}{PARAM_DESCRIPTION, Default: 'gradient'}

\item{palette}{PARAM_DESCRIPTION, Default: NULL}

\item{no_labels}{PARAM_DESCRIPTION, Default: TRUE}

\item{title}{PARAM_DESCRIPTION, Default: NULL}

\item{subtitle}{PARAM_DESCRIPTION, Default: NULL}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
 #EXAMPLE1
}
}
\author{
Lorenzo Busetto, phD (2017) \href{mailto:lbusett@gmail.com}{lbusett@gmail.com}
}
