% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_rast_gg.R
\name{plot_rast_gg}
\alias{plot_rast_gg}
\title{plot_rast_gg}
\usage{
plot_rast_gg(in_rast, xlims = NULL, ylims = NULL, basemap = NULL,
  zoomin = -1, scalebar = TRUE, scalebar_dist = NULL,
  na.color = "transparent", na.value = NA, palette_type = "gradient",
  palette = NULL, labels = NULL, no_labels = TRUE, title = NULL,
  subtitle = NULL)
}
\arguments{
\item{in_rast}{PARAM_DESCRIPTION}

\item{xlims}{PARAM_DESCRIPTION, Default: NULL}

\item{ylims}{PARAM_DESCRIPTION, Default: NULL}

\item{basemap}{PARAM_DESCRIPTION, Default: NULL}

\item{zoomin}{PARAM_DESCRIPTION, Default: -1}

\item{scalebar}{PARAM_DESCRIPTION, Default: TRUE}

\item{scalebar_dist}{PARAM_DESCRIPTION, Default: NULL}

\item{na.color}{PARAM_DESCRIPTION, Default: 'transparent'}

\item{na.value}{PARAM_DESCRIPTION, Default: NA}

\item{palette_type}{PARAM_DESCRIPTION, Default: 'gradient'}

\item{palette}{PARAM_DESCRIPTION, Default: NULL}

\item{labels}{PARAM_DESCRIPTION, Default: NULL}

\item{no_labels}{PARAM_DESCRIPTION, Default: TRUE}

\item{title}{PARAM_DESCRIPTION, Default: NULL}

\item{subtitle}{PARAM_DESCRIPTION, Default: NULL}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
 in_rast <- raster::stack(system.file("extdata/OLI_test", "oli_multi_1000_b2.tif",
                                  package = "sprawl.data"))
 plot_rast_gg(in_rast, basemap = "cartolight",
                  palette_type = "diverging",
                  no_labels = T,
                  na.value = 0,
                  zoomin = 0, title = "OLI", subtitle = "Band 2")

 in_rast <- system.file("extdata/REYE_test", "REYE_2016_185_gNDVI.tif",
                          package = "sprawl.data")
 plot_rast_gg(in_rast, basemap = "cartolight",
                  palette_type = "diverging",
                  no_labels = F,
                  na.value = 0,
                  zoomin = 0,
                  title = "Rapideye", subtitle = "gNDVI")
}
}
\author{
Lorenzo Busetto, phD (2017) \href{mailto:lbusett@gmail.com}{lbusett@gmail.com}
}
