% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_vect.R
\name{plot_vect}
\alias{plot_vect}
\title{FUNCTION_TITLE}
\usage{
plot_vect(in_data, line_color = "black", line_size = 0.2, fill_var = NULL,
  transparency = 0, facet_var = NULL, facet_rows = NULL,
  borders_layer = NULL, borders_color = "grey15", borders_size = 0.2,
  borders_txt_field = NULL, borders_txt_size = 2.5,
  borders_txt_color = "grey15", basemap = NULL, zoomin = 0,
  xlims = NULL, ylims = NULL, zlims = NULL, zlims_type = "vals",
  outliers_style = "recolor", outliers_colors = c("grey10", "grey90"),
  scalebar = FALSE, scalebar_dist = NULL, na.color = NULL,
  na.value = NULL, palette_name = NULL, direction = 1, leg_type = NULL,
  leg_labels = NULL, leg_breaks = NULL, leg_position = "right",
  no_axis = FALSE, title = "Vector Plot", subtitle = NULL,
  theme = theme_bw(), grid = FALSE, verbose = TRUE)
}
\arguments{
\item{in_data}{PARAM_DESCRIPTION}

\item{line_color}{\code{character} color used to plot the polygon borders of
in_data}

\item{line_size}{\code{numeric} size of lines used to plot the polygons borders of
in_data, Default: 0.2}

\item{fill_var}{PARAM_DESCRIPTION, Default: NULL}

\item{transparency}{PARAM_DESCRIPTION, Default: 0}

\item{facet_var}{PARAM_DESCRIPTION, Default: NULL}

\item{facet_rows}{PARAM_DESCRIPTION, Default: NULL}

\item{borders_layer}{PARAM_DESCRIPTION, Default: NULL}

\item{borders_color}{PARAM_DESCRIPTION, Default: 'grey15'}

\item{borders_size}{PARAM_DESCRIPTION, Default: 0.2}

\item{borders_txt_field}{PARAM_DESCRIPTION, Default: NULL}

\item{borders_txt_size}{PARAM_DESCRIPTION, Default: 1}

\item{borders_txt_color}{PARAM_DESCRIPTION, Default: 'grey15'}

\item{basemap}{PARAM_DESCRIPTION, Default: NULL}

\item{zoomin}{PARAM_DESCRIPTION, Default: 0}

\item{xlims}{PARAM_DESCRIPTION, Default: NULL}

\item{ylims}{PARAM_DESCRIPTION, Default: NULL}

\item{zlims}{PARAM_DESCRIPTION, Default: NULL}

\item{zlims_type}{PARAM_DESCRIPTION, Default: 'vals'}

\item{outliers_style}{PARAM_DESCRIPTION, Default: 'recolor'}

\item{outliers_colors}{PARAM_DESCRIPTION, Default: c("grey10", "grey90")}

\item{scalebar}{PARAM_DESCRIPTION, Default: FALSE}

\item{scalebar_dist}{PARAM_DESCRIPTION, Default: NULL}

\item{na.color}{PARAM_DESCRIPTION, Default: NULL}

\item{na.value}{PARAM_DESCRIPTION, Default: NULL}

\item{palette_name}{PARAM_DESCRIPTION, Default: NULL}

\item{direction}{PARAM_DESCRIPTION, Default: 1}

\item{leg_type}{PARAM_DESCRIPTION, Default: NULL}

\item{leg_labels}{PARAM_DESCRIPTION, Default: NULL}

\item{leg_breaks}{PARAM_DESCRIPTION, Default: NULL}

\item{leg_position}{PARAM_DESCRIPTION, Default: 'right'}

\item{no_axis}{PARAM_DESCRIPTION, Default: FALSE}

\item{title}{PARAM_DESCRIPTION, Default: 'Vector Plot'}

\item{subtitle}{PARAM_DESCRIPTION, Default: NULL}

\item{theme}{PARAM_DESCRIPTION, Default: theme_bw()}

\item{grid}{PARAM_DESCRIPTION, Default: FALSE}

\item{verbose}{PARAM_DESCRIPTION, Default: TRUE}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 library(ggplot2)
 in_vect <- get(load(system.file("extdata/shapes", "poly_lomb.RData",
                                 package = "sprawl.data")))
 # plot only geometry
 plot_vect(in_vect)
 plot_vect(in_vect, line_color = "blue", line_size = 1.5)

 # plot with a fill on a cartegorical variable with a single "level"
 plot_vect(in_vect, fill_var = "NAME_2")

 # change the palette, add a scalebar and remove the grid
 plot_vect(in_vect, fill_var = "NAME_2",no_axis = F, palette_name = "Set3",
           scalebar = TRUE, grid = FALSE)


 # plot with a fill on a continuous variable with two "levels", using facets
 # and a diverging palette. also add a "borders" layer with a different
 #plot color
 plot_vect(in_vect, fill_var = "population", facet_var = "year",
           palette = "RdYlBu", scalebar = T, scalebar_dist = 50,
           grid = FALSE, zlims = c(5,20), outliers_colors = c("yellow", "green"),
           borders_layer = get_boundaries("ITA", level = 0),
           borders_color = "red")

 }
}
}
\author{
Lorenzo Busetto, phD (2017) \href{mailto:lbusett@gmail.com}{lbusett@gmail.com}
}
