% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_vect.R
\name{plot_vect}
\alias{plot_vect}
\title{FUNCTION_TITLE}
\usage{
plot_vect(in_data, fill_var = NULL, fill_transparency = 0,
  polys_color = "black", polys_size = 0.2, facet_var = NULL,
  facet_rows = NULL, borders_layer = NULL, borders_color = "grey15",
  borders_size = 0.2, borders_txt_fiels = NULL, borders_txt_size = 1,
  borders_txt_color = "grey15", basemap = NULL, zoomin = 0,
  xlims = NULL, ylims = NULL, zlims = NULL, zlims_type = "vals",
  outliers_style = "recolor", outliers_colors = c("grey10", "grey90"),
  scalebar = FALSE, scalebar_dist = NULL, na.color = NULL,
  na.value = NULL, palette_name = "Greens", direction = 1,
  leg_type = NULL, leg_labels = NULL, leg_breaks = NULL,
  leg_position = "right", no_axis = FALSE, title = "Vector Plot",
  subtitle = NULL, theme = theme_bw(), verbose = TRUE)
}
\arguments{
\item{in_data}{PARAM_DESCRIPTION}

\item{fill_var}{PARAM_DESCRIPTION}

\item{facet_var}{PARAM_DESCRIPTION}

\item{facet_rows}{PARAM_DESCRIPTION, Default: NULL}

\item{borders_layer}{PARAM_DESCRIPTION, Default: NULL}

\item{borders_color}{PARAM_DESCRIPTION, Default: 'grey15'}

\item{borders_size}{PARAM_DESCRIPTION, Default: 0.2}

\item{borders_txt_fiels}{PARAM_DESCRIPTION, Default: NULL}

\item{borders_txt_size}{PARAM_DESCRIPTION, Default: 1}

\item{borders_txt_color}{PARAM_DESCRIPTION, Default: 'grey15'}

\item{basemap}{Yet to be implemented !!!!!, Default: NULL}

\item{zoomin}{PARAM_DESCRIPTION, Default: 0}

\item{xlims}{PARAM_DESCRIPTION, Default: NULL}

\item{ylims}{PARAM_DESCRIPTION, Default: NULL}

\item{zlims}{PARAM_DESCRIPTION, Default: NULL}

\item{zlims_type}{PARAM_DESCRIPTION, Default: 'vals'}

\item{outliers_style}{PARAM_DESCRIPTION, Default: 'recolor'}

\item{outliers_colors}{PARAM_DESCRIPTION, Default: c("grey10", "grey90")}

\item{scalebar}{PARAM_DESCRIPTION, Default: FALSE}

\item{scalebar_dist}{PARAM_DESCRIPTION, Default: NULL}

\item{na.color}{PARAM_DESCRIPTION, Default: NULL}

\item{na.value}{PARAM_DESCRIPTION, Default: NULL}

\item{palette_name}{PARAM_DESCRIPTION, Default: NULL}

\item{direction}{PARAM_DESCRIPTION, Default: 1}

\item{leg_type}{PARAM_DESCRIPTION, Default: NULL}

\item{leg_labels}{PARAM_DESCRIPTION, Default: NULL}

\item{leg_breaks}{PARAM_DESCRIPTION, Default: NULL}

\item{no_axis}{PARAM_DESCRIPTION, Default: FALSE}

\item{title}{PARAM_DESCRIPTION, Default: 'Vector Plot'}

\item{subtitle}{PARAM_DESCRIPTION, Default: NULL}

\item{theme}{PARAM_DESCRIPTION, Default: theme_bw()}

\item{verbose}{PARAM_DESCRIPTION, Default: TRUE}

\item{palette_type}{PARAM_DESCRIPTION, Default: 'gradient'}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 in_vect <- get_boundaries("ITA", level = 3) \%>\%
    dplyr::filter(NAME_1 == "Lombardia")
 geom <- in_vect \%>\%
   select(NAME_1, NAME_2, NAME_3)
 st_geometry(in_vect) <- NULL
 in_data <- in_vect  \%>\%
    dplyr::select(NAME_1, NAME_2, NAME_3) \%>\%
    dplyr::mutate(y_2003 = sample(dim(.)[1]),
                  y_2016 = 3*sample(dim(.)[1])) \%>\%
    tidyr::gather(year, population, -NAME_1, -NAME_2, -NAME_3)
 in_vect <- dplyr::left_join(in_data, geom) \%>\%
   sf::st_as_sf()
 plot_vect(in_vect,
           fill_var = "population",
           facet_var = "year",
           scalebar = TRUE,
           scalebar_dist = 30,
           theme = theme_light())
 }
}
}
\author{
Lorenzo Busetto, phD (2017) \href{mailto:lbusett@gmail.com}{lbusett@gmail.com}
}
