% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_rast.R
\name{read_rast}
\alias{read_rast}
\title{read a raster file from disk}
\usage{
read_rast(in_file, bands = NULL, verbose = TRUE)
}
\arguments{
\item{in_file}{\code{character} filename to be read}

\item{bands}{\code{numeric array} If the input is a multiband raster,
band numbers to be read (e.g., bands = c(1,5,7) will read only the
specified bands in a \code{RasterStack})}

\item{verbose}{\code{logical} If FALSE, processing mesasges are suppressed,
Default: TRUE}
}
\value{
a \code{RasterLayer}, \code{RasterStack} or \code{RasterBrick} object
}
\description{
reads a raster file to a \code{RasterLayer}, \code{RasterStack} or
\code{RasterBrick}, automatically checking if the input file is valid and is
a single- or multi-band object.
}
\details{
DETAILS
}
\examples{
\dontrun{
 # read a singleband file
  in_file <- system.file("extdata/REYE_test", "REYE_2016_185_gNDVI.tif",
                          package = "sprawl.data")
  read_rast(in_file)

  # read a multiband file
  in_file <- system.file("extdata/MODIS_test", "EVIts_test.tif",
                          package = "sprawl.data")
  read_rast(in_file)

  # read a multiband file selecting specific bands
  read_rast(in_file, bands = c(1,10))

}
}
\author{
Lorenzo Busetto, phD (2017) \href{mailto:lbusett@gmail.com}{lbusett@gmail.com}
}
