% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reproj_extent.R
\name{reproj_extent}
\alias{reproj_extent}
\title{reproject the extent of a spatial object}
\usage{
reproj_extent(ext_in, in_proj = NULL, out_proj = NULL)
}
\arguments{
\item{ext_in}{object of class \code{\link[sprawl:sprawlext]{sprawl::sprawlext}}, or any other object or
filename from which a \code{sprawlext} object can be derived}

\item{in_proj}{\code{character} proj4string representing the projection of the
input extent}

\item{out_proj}{\code{character} proj4string representing the desired projection
for the output extent}
}
\value{
reprojected extent (in out_proj projection)
}
\description{
accessory function to convert the extent of a spatial object to
a different projection. simple wrapper for the \code{spTransform} function on
a \code{sprawlext} object
}
\examples{
{
 \dontrun{
 EXAMPLE_1
}
}
}
