% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_extent.R
\name{select_extent}
\alias{select_extent}
\title{Interactively select an extent}
\usage{
select_extent(in_object = NULL, reproject = TRUE, transparency = 0.5)
}
\arguments{
\item{in_object}{a \code{R} spatial object of any kind supported by
\code{mapview} ("Raster", "sp", "sf") to be shown as reference on the map
(see examples). If NULL, an empty map of the world is shown, Default: NULL}

\item{reproject}{\code{logical} If TRUE, and a valid \code{in_object} was provided,
the selected extent is returned in the same CRS of \code{in_object}, otherwise
it is returned in CRS 4326}

\item{transparency}{PARAM_DESCRIPTION, Default: 0.5}
}
\value{
a \code{sprawlext} object representing the selected extent.
}
\description{
Select an extent interactively from a leaflet map
}
\details{
DETAILS
}
\examples{
\donttun{
if(interactive()){
 in_rast <- read_rast(system.file("extdata/REYE_test", "REYE_2016_185_gNDVI.tif",
                          package = "sprawl.data"))
 extent    <- select_extent(in_rast)
 extent
 }
 }
}
\author{
Lorenzo Busetto, phD (2017) \href{mailto:lbusett@gmail.com}{lbusett@gmail.com}
}
