% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{sprawlext-class}
\alias{sprawlext-class}
\alias{setClasses}
\title{helper to set classes on package loading}
\usage{
setClasses()
}
\value{
OUTPUT_DESCRIPTION
}
\description{
helper to set classes on package loading

An S4 class to represent the extent of a spatial object,
associated with its proj4
string.
}
\details{
DETAILS
}
\section{Slots}{

\describe{
\item{\code{extent}}{\code{numeric (4)} extent of the object (xmin, ymin, xmax, ymax)}

\item{\code{proj4string}}{character` proj4string of the object}
}}

