% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sprawl_scalebar.R
\name{sprawl_scalebar}
\alias{sprawl_scalebar}
\title{sprawl_scalebar}
\usage{
sprawl_scalebar(in_obj, location = "bottomright", scalebar_dist = NULL,
  height = 0.02, st.dist = 0.01, st.bottom = TRUE, st.size = 3.5,
  st.color = "black", box.fill = c("black", "white"),
  box.color = "black", dd2km = FALSE, model, x.min, x.max, y.min,
  y.max, anchor = NULL, facet.var = NULL, facet.lev = NULL,
  units = "m")
}
\arguments{
\item{in_obj}{\code{Raster} object passed to \code{plot_rast_gg} or \code{plot_vect} to which
the scalebar should be added;}

\item{location}{\code{character ["bottomright" | "bottomleft" | "topleft" | "topright"]}
indicating where the scalebar should be placed, Default: "bottomright"}

\item{scalebar_dist}{\code{numeric} distance in to represent with each segment of the scale bar.}

\item{height}{\code{numeric} value between 0 and 1 to indicate the height of the
scale bar, as a proportion of the y axis, Default: 0.02}

\item{st.dist}{\code{numeric} value between 0 and 1 to indicate the distance
between the scale bar and the scale bar text, as a proportion of the y axis,
Default: 0.02}

\item{st.bottom}{\code{logical} If TRUE (default) the scale bar text is displayed
at the bottom of the scale bar, if FALSE, it is displayed at the top.}

\item{st.size}{\code{numeric} value to indicate the scale bar text size. It is
passed to the size argument of \code{ggplot::annotate} function.}

\item{st.color}{Text color for the scalebar, Default: 'black'}

\item{box.fill}{fill colors for the scalebar polygons,
Default: c("black", "white")}

\item{box.color}{line color for the scalebar polygons, Default: 'black'}

\item{dd2km}{\code{logical} If TRUE it is assumed that map coordinates are in
decimal degrees, if FALSE, it assumed they are in meters, Default: FALSE}

\item{model}{\code{character} choice of ellipsoid model ("WGS84", "GRS80", "Airy",
"International", "Clarke", "GRS67"). Ignored if dd2km is FALSE.}

\item{x.min}{if \code{data} is not defined, number with the minimum x coordinate.}

\item{x.max}{if \code{data} is not defined, number with the maximum x coordinate.}

\item{y.min}{if \code{data} is not defined, number with the minimum y coordinate.}

\item{y.max}{if \code{data} is not defined, number with the maximum y coordinate.}

\item{anchor}{\code{named numeric(2)} with coordinates to control the
symbol position. For \code{location = "topright"}, \code{anchor}
defines the coordinates of the symbol's topright corner and so forth.
The x coordinate must be named as x and the y coordinate as y.}

\item{facet.var}{if faceting, character vector of variable names used for
faceting. This is useful for placing the scalebar only in one facet and must
be used together with \code{facet.lev}.}

\item{facet.lev}{character vector with the name of one level for each
variable in \code{facet.var}. The scale bar will be drawn only in the \code{facet.lev}
facet.}

\item{units}{DESCRIPTION}
}
\description{
Adds a scale bar to maps created with \code{plot_rast_gg}.
}
\details{
Simple modifications on the \code{scalebar} function in package \code{ggsn}
}
\author{
\itemize{
\item Original code by Oswaldo Santos Baquero - ggsn: North Symbols and
Scale Bars for Maps Created with 'ggplot2' or 'ggmap'. R package version
0.4.3. https://github.com/oswaldosantos/ggsn
\item Slight modifications to adapt it to \code{sprawl} needs by Lorenzo Busetto:
\itemize{
\item Removed measure units text from all but the last numeric label;
\item Added support for non-metric (e.g., feet) projections by adding the \code{units}
argument and changing the text accordingly;
\item fontface to bold in the scalebar to improve visibility;
\item automatic computation of distance if \code{dist} is not passed
}
}
}
