% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fast_summ.R, R/summarize_table.R
\name{summarize_data}
\alias{summarize_data}
\alias{summarize_data}
\title{extract useful statistics from a data.table (or data.frame or tibble)}
\usage{
summarize_data(in_data, var, comp_quant, FUN, band_n, selband)

summarize_data(in_data, var, comp_quant, FUN, band_n, selband)
}
\arguments{
\item{in_data}{PARAM_DESCRIPTION}

\item{var}{PARAM_DESCRIPTION}

\item{comp_quant}{PARAM_DESCRIPTION}

\item{FUN}{PARAM_DESCRIPTION}

\item{band_n}{PARAM_DESCRIPTION}

\item{selband}{PARAM_DESCRIPTION}

\item{in_data}{PARAM_DESCRIPTION}

\item{var}{PARAM_DESCRIPTION}

\item{comp_quant}{PARAM_DESCRIPTION}

\item{FUN}{PARAM_DESCRIPTION}

\item{band_n}{PARAM_DESCRIPTION}

\item{selband}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION

OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION

FUNCTION_DESCRIPTION
}
