% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_vector_to_pgres.R
\name{write_vector_to_pgres}
\alias{write_vector_to_pgres}
\title{write a spatial table to a PostGIS db table}
\usage{
write_vector_to_pgres(in_vect, con, table_name, schema = "public",
  make_valid = FALSE, drop = FALSE, update = FALSE, row.names = FALSE,
  verbose = TRUE)
}
\arguments{
\item{in_vect}{either an "sf" or "sp" R object, or the full path of
a valid vector file}

\item{con}{a \code{PostgreSQLConnection} object to a PostGIS DB, created by a call to
\code{DBI::DBI::dbConnect(PostgreSQL(),....}}

\item{table_name}{\code{character} name of the table to be used for storage in the db}

\item{schema}{\code{character} name of the schema to be used for storage in the db}

\item{make_valid}{\code{logical} If FALSE, the function aborts in case invalid
geometries are found. If TRUE, an automatic correction based on
\code{lwgeom::st_make_valid} is attempted, Default: FALSE}

\item{drop}{\code{logical} indicating if existing tables should be overwrittern,
Default: FALSE}

\item{row.names}{\code{logical} indicating if an additional row containing row numbers,
should be added to the DB table, Default: FALSE}

\item{drop}{\code{logical} indicating if data should be appended to existing tables,
Default: FALSE}
}
\value{
The function is called for its side effects
}
\description{
write spatial table to a PostGIS db table (simple wrapper around
\code{sf::st_write_db} with automatic casting to \code{sf} objects and argument
checks)
}
\examples{
\dontrun{
# establish a connection with a postGIS DB
con <- DBI::dbConnect(PostgreSQL(),
                     host = "xx.xxx.xx.xx", dbname = "xxxxx", user = "xxxxx",
                     password = "xxxxx", port = 5432)
> con
<PostgreSQLConnection>

 # write a vector object in a table of the DB
 write_vector_to_pgres(in_vect,
                  con        = con,
                  table_name = "my_table",
                  schema     = "public",
                  drop       = TRUE)
}
}
\seealso{
\code{sf::st_write_db} \code{lwgeom::st_make_valid} \code{sf::st_is_valid}
}
\author{
Lorenzo Busetto, phD (2017) \href{mailto:lbusett@gmail.com}{lbusett@gmail.com}
}
